% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortDB_comp.r
\name{BNDBfrequency}
\alias{BNDBfrequency}
\title{Calculates the internal frequencies of BNDB cohorts}
\usage{
BNDBfrequency(
  internalBNDB,
  smappath,
  smap,
  buildBNInternalDB = FALSE,
  smapdata,
  input_fmt_SV = c("Text", "dataFrame"),
  dbOutput = c("dataframe", "text"),
  BNDBpath,
  BNDBpattern,
  outpath,
  win_indel = 10000,
  win_inv_trans = 50000,
  perc_similarity = 0.5,
  indelconf = 0.5,
  invconf = 0.01,
  limsize = 1000,
  transconf = 0.1,
  returnMethod = c("Text", "dataFrame"),
  EnzymeType = c("SVmerge", "SE")
)
}
\arguments{
\item{internalBNDB}{character. Path to the merged SV files.}

\item{smappath}{character. path to the query smap file.}

\item{smap}{character. File name for the smap}

\item{buildBNInternalDB}{boolean. Checking whether the merged BNDB 
file database exist.}

\item{smapdata}{dataframe. smapdata in the form of dataframe.}

\item{input_fmt_SV}{character. Choice between Text and DataFrame.}

\item{dbOutput}{character. database output type.
Options dataframe or text.}

\item{BNDBpath}{character. Path to the BNDB file database.}

\item{BNDBpattern}{character. pattern of the file names to merge.}

\item{outpath}{character. Path to merged SV solo datasets.}

\item{win_indel}{Numeric. Insertion and deletion error window.}

\item{win_inv_trans}{Numeric. Inversion and translocation error window.}

\item{perc_similarity}{Numeric . ThresholdPercentage similarity 
of the query SV and reference SV.}

\item{indelconf}{Numeric. Threshold for insertion and deletion Score.}

\item{invconf}{Numeric. Threshold for inversion Score.}

\item{limsize}{Numeric. SV size limit.}

\item{transconf}{Numeric. Threshold for translocation Score.}

\item{returnMethod}{character. Choice between Text and DataFrame.}

\item{EnzymeType}{Character. Type of enzyme. Options SVmerge and SE.}
}
\value{
Text file or data frames containing internalFrequency data.
}
\description{
Calculates the internal frequencies of BNDB cohorts
}
\examples{
path <- system.file("extdata", "Bionano_config/", package = "nanotatoR")
pattern <- "*_hg19_*"
smapName="GM24385_Ason_DLE1_VAP_trio5.smap"
smap = system.file("extdata", smapName, package="nanotatoR")
BNDBfrequency(smap = smap, 
 buildBNInternalDB=TRUE, 
   input_fmt_SV = "Text",
   dbOutput="dataframe",
   BNDBpath = path, 
   BNDBpattern = pattern, 
   outpath, 
   win_indel = 10000,
   win_inv_trans = 50000, 
 perc_similarity = 0.5, 
   indelconf = 0.5, 
   invconf = 0.01, 
   limsize = 1000,
 transconf = 0.1,
   returnMethod=c("dataFrame"), 
   EnzymeType = c("SE"))
}
