% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/OverlappingGenes.r
=======
% Please edit documentation in R/Bed_SV_Comp.r
>>>>>>> 714868ed579a092f4cd6df397baa7f46eded618f
\name{buildrunBNBedFiles}
\alias{buildrunBNBedFiles}
\title{Reads BED files to produce bionano Bed files}
\usage{
<<<<<<< HEAD
buildrunBNBedFiles(
  bedFile,
  returnMethod = c("Text", "dataFrame"),
  outdir,
  fname
)
=======
buildrunBNBedFiles(bedFile, returnMethod = c("Text", "dataFrame"),
  outdir)
>>>>>>> 714868ed579a092f4cd6df397baa7f46eded618f
}
\arguments{
\item{bedFile}{character. Path to UCSC Bed File.}

<<<<<<< HEAD
\item{returnMethod}{character. Path to output directory.}

\item{outdir}{character. Path to output directory.}

\item{fname}{character. Output File name.}
=======
\item{outdir}{character. Path to output directory.}
>>>>>>> 714868ed579a092f4cd6df397baa7f46eded618f
}
\value{
Data Frame or text file. Contains the gene information.
}
\description{
Reads BED files to produce bionano Bed files
}
\examples{
<<<<<<< HEAD
bedFile <- system.file("extdata", "HomoSapienGRCH19_lift37.bed", 
      package="nanotatoR")
bed<-buildrunBNBedFiles(bedFile,returnMethod="dataFrame")
=======
\dontrun{
bedFile = "C:\\\\Annotator\\\\Data\\\\Homo_sapiens.Hg19.bed"
buildrunBNBedFiles(bedFile,outdir=tempdir())
}
>>>>>>> 714868ed579a092f4cd6df397baa7f46eded618f
}
