% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortDB_comp.r
\name{cohortFrequency}
\alias{cohortFrequency}
\title{Calculates the internal frequencies of SV in bionano cohorts}
\usage{
cohortFrequency(internalBNDB, smappath, smap, buildBNInternalDB = FALSE,
  smapdata, input_fmt = c("Text", "dataFrame"),
  dbOutput = c("dataframe", "text"), BNDBPath, BNDBPattern, fname,
  outpath, win_indel = 10000, win_inv_trans = 50000,
  perc_similarity = 0.5, indelconf = 0.5, invconf = 0.01,
  limsize = 1000, transconf = 0.1, returnMethod = c("Text",
  "dataFrame"))
}
\arguments{
\item{internalBNDB}{character. Path to the merged BN files.}

\item{smappath}{character. path to the query smap file.}

\item{smap}{character. File name for the smap}

\item{buildBNInternalDB}{boolean. Checking whether the merged bionano 
file database exist.}

\item{smapdata}{character. SV data in dataframe.}

\item{input_fmt}{character. Choice between Text and DataFrame.}

\item{dbOutput}{character. Output of merged bionano data.}

\item{BNDBPath}{Path of Bionano database files.}

\item{BNDBPattern}{Pattern of Bionano database files.}

\item{fname}{character. Filename in case dbOutput = Text.}

\item{outpath}{character. Path to merged SV solo datasets.}

\item{win_indel}{Numeric. Insertion and deletion error window.}

\item{win_inv_trans}{Numeric. Inversion and translocation error window.}

\item{perc_similarity}{Numeric . ThresholdPercentage similarity 
of the query SV and reference SV.}

\item{indelconf}{Numeric. Threshold for insertion and deletion Score.}

\item{invconf}{Numeric. Threshold for inversion Score.}

\item{limsize}{Numeric . Minimum size to consider for a SV.}

\item{transconf}{Numeric. Threshold for translocation Score.}

\item{returnMethod}{character. Choice between Text and DataFrame.}
}
\value{
Text file or data frames containing internalFrequency data.
}
\description{
Calculates the internal frequencies of SV in bionano cohorts
}
\examples{
mergedFiles <- system.file("extdata", "BNSOLO2_merged.txt", 
package = "nanotatoR")
smapName <- "F1.1_TestSample1_solo_hg19.smap"
smappath <- system.file("extdata",  package = "nanotatoR")
win_indel = 10000; win_inv_trans = 50000; perc_similarity = 0.5;
indelconf = 0.5; invconf = 0.01;transconf = 0.1
cohortFreq<-cohortFrequency(internalBNDB = mergedFiles , smappath , 
smap=smapName, input_fmt ="Text", 
buildBNInternalDB=FALSE, win_indel, win_inv_trans, 
perc_similarity , indelconf, invconf, 
transconf,returnMethod="dataFrame")
}
