% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DecipherExtraction.r
\name{Decipherfrequency}
\alias{Decipherfrequency}
\title{Frequency calculation of variants compared to Decipher.}
\usage{
Decipherfrequency(
  decipherpath,
  smap,
  smap_data,
  win_indel = 10000,
  perc_similarity = 0.5,
  returnMethod = c("Text", "dataFrame"),
  input_fmt_SV = c("Text", "dataFrame"),
  EnzymeType = c("SVMerge", "SE"),
  outpath
)
}
\arguments{
\item{decipherpath}{character. Decipher Text file.}

\item{smap}{character Filepath for smap.}

\item{smap_data}{Dataset containing smap data.}

\item{win_indel}{character indel window. Default 10000.}

\item{perc_similarity}{Numeric . ThresholdPercentage similarity 
of the query SV and reference SV.}

\item{returnMethod}{character. Choice between 
text or data frame as the output.}

\item{input_fmt_SV}{boolean . Options SE and SVMerge.}

\item{EnzymeType}{boolean . Options SE and SVMerge.}

\item{outpath}{character. Path where gene lists are saved.}
}
\value{
dataframe containing decipher data.
are stored as text files.
}
\description{
Frequency calculation of variants compared to Decipher.
}
\examples{
decipherpath = system.file("extdata", "population_cnv.txt", package="nanotatoR")
smappath=system.file("extdata", "GM24385_Ason_DLE1_VAP_trio5.smap", package="nanotatoR")
datdecipher <- Decipherfrequency (decipherpath = decipherpath, 
smap = smappath, win_indel = 10000,
EnzymeType= "SE",
perc_similarity = 0.5,returnMethod="dataFrame", 
input_fmt_SV = "Text")
datdecipher[1,]
}
