% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/DGV_extraction_WIP.r
=======
% Please edit documentation in R/DGV_extraction.r
>>>>>>> 04c1b43bc3aded9fbd47aace30093a5b16389a06
\name{DGV_extraction}
\alias{DGV_extraction}
\title{Frequency calculation of variants compared to DGV.}
\usage{
<<<<<<< HEAD
DGV_extraction(
  hgpath,
  smappath,
  smap,
  smap_data,
  input_fmt_DGV = c("Text", "DataFrame"),
  win_indel_DGV = 10000,
  win_inv_trans_DGV = 50000,
  perc_similarity_DGV = 0.5,
  returnMethod = c("Text", "dataFrame"),
  outpath
)
=======
DGV_extraction(hgpath, smappath, smap, smap_data,
  input_fmt_DGV = c("Text", "dataFrame"), win_indel_DGV = 10000,
  win_inv_trans_DGV = 50000, perc_similarity_DGV = 0.5,
  returnMethod = c("Text", "dataFrame"), outpath, usample = 54946)
>>>>>>> 04c1b43bc3aded9fbd47aace30093a5b16389a06
}
\arguments{
\item{hgpath}{character. Path to Database of Genomic Variants (DGV)
Text file.}

\item{smappath}{character. Path for smap textfile.}

\item{smap}{character. File name for smap textfile.}

\item{smap_data}{dataframe. Dataset containing smap data.}

\item{input_fmt_DGV}{character. Choice between text or data frame as
an input to the DGV frequency calculator.}

<<<<<<< HEAD
\item{win_indel_DGV}{Numeric. Insertion and deletion error window.Default 10000
=======
\item{win_indel_DGV}{Numeric. Insertion and deletion error window.Default 
10000
>>>>>>> 04c1b43bc3aded9fbd47aace30093a5b16389a06
bases.}

\item{win_inv_trans_DGV}{Numeric. Inversion and translocation error window.
Default 50000 bases.}

<<<<<<< HEAD
\item{perc_similarity_DGV}{Numeric . ThresholdPercentage similarity 
of the query SV and reference SV. Default 0.5.}

\item{returnMethod}{character. Choice between text or data frame as the output.}

\item{outpath}{character. Path where gene lists are saved.}
}
\value{
Text and character vector containg gene list and terms associated with them
=======
\item{perc_similarity_DGV}{Numeric. ThresholdPercentage similarity 
of the query SV and reference SV. Default 0.5.}

\item{returnMethod}{character. Choice between text or data frame as the
output.}

\item{outpath}{character. Path where gene lists are saved.}

\item{usample}{Numeric. Number of unique samples.}
}
\value{
Text and character vector containg gene list and terms associated 
with them
>>>>>>> 04c1b43bc3aded9fbd47aace30093a5b16389a06
        are stored as text files.
}
\description{
Frequency calculation of variants compared to DGV.
}
\examples{
\dontrun{
<<<<<<< HEAD
hgpath="C:\\\\nanotatoR\\\\Data\\\\GRCh37_hg19_variants_2016-05-15.txt";
smappath="Z:/Hayk's_Materials/Bionano/Projects/UDN/F1_UDN287643_Benic.Aria/";
smap="F1_UDN287643_P_Q.S_VAP_SVmerge_trio_access.txt";
win_indel_DGV=10000;win_inv_trans_DGV=50000;perc_similarity_DGV=0.5
DGV_extraction (hgpath, smappath, win_indel_DGV = 10000, win_inv_trans_DGV = 50000,
perc_similarity_DGV = 0.5,returnMethod="dataFrame")
=======
smap <- "F1.1_TestSample1_solo_hg19.smap"
smappath <- system.file("extdata",  package = "nanotatoR")
hgpath <- system.file("extdata", "GRCh37_hg19_variants_2016-05-15.txt", 
package = "nanotatoR")
win_indel_DGV=10000;win_inv_trans_DGV=50000;perc_similarity_DGV=0.5;
usample = 54946
dgv <- DGV_extraction (hgpath, input_fmt_DGV = "Text",smap=smap, 
smappath, win_indel_DGV = 10000, win_inv_trans_DGV = 50000,
perc_similarity_DGV = 0.5,returnMethod="dataFrame",usample = 54946)
>>>>>>> 04c1b43bc3aded9fbd47aace30093a5b16389a06
}
}
