% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractclinvar_mod.r
\name{extract_clinvar_mod}
\alias{extract_clinvar_mod}
\title{Extract the genes and variants related to a genetic disorder from ClinVar}
\usage{
extract_clinvar_mod(
  keyword,
  localPDB.path,
  type = "both",
  HPO.disease = NULL,
  genelist = NULL,
  OMIM = NULL
)
}
\arguments{
\item{keyword}{character. character string: keyword,
to search a disease, a clinical feature, or a phenotype.}

\item{localPDB.path}{character. the path of
localized public data bases. The default value is set in the working
directory.}

\item{type}{character. the type of the information to extract,
must be one of "gene", "variant","both"(default).}

\item{HPO.disease}{character. MIM number of the disease.
The default value is NULL, which means that all the
OMIM number of the disease in HPO are added.
localized public data bases. The default value is set in the working
directory.}

\item{genelist}{character. the gene(s) associated
to the disease, or the genes you are interested.}

\item{OMIM}{character. whether use the information from
OMIM database. The default value is NULL.
It can be set 'yes' when you make sue you have a OMIM API key.}
}
\value{
subset of the file gene_condition_source_id, which include
all the information about genes and phenotypes in ClinVar and
subset of the file variant_summary.txt,
but added sevetal colomns which describe the phenotype
from GeneReview, MedGen, and OMIM databases. Function modified from
extract_clinvar function VarFromPDB.
}
\description{
Extract the genes and variants related to a genetic disorder from ClinVar
}
\examples{
keyword = "retinoblastoma"
extract_clinvar_mod(keyword,
localPDB.path = system.file("extdata", "localPDB", package="nanotatoR"),
type = "both", HPO.disease = NULL,
genelist = NULL, OMIM = NULL)
}
