% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergesmap_SVMerge_SE.R
\name{FamilyInfoPrep}
\alias{FamilyInfoPrep}
\title{Mapping Realtionship to unique nanoIDs}
\usage{
FamilyInfoPrep(
  Samplecodes = "X:/Hayks_Materials/BNG/Projects/nanotatoR_sample_codes.csv",
  mergeKey = "X:/Hayks_Materials/BNG/Projects/MergeKey.csv",
  outMode = c("Text", "dataframe"),
  outpath = "X:/Hayks_Materials/BNG/Projects/VAP_DLE1_solo_SMAPs/Merged"
)
}
\arguments{
\item{Samplecodes}{character. File containing relations and IDs associated to them.}

\item{mergeKey}{character. File containing sample ID and relation.}

\item{outMode}{character. The ouput mode. Choices, dataframe or Text.}

\item{outpath}{character. Path where the dual labelled merged samples are kept.
Is mandatory if outMode is Text.}
}
\value{
Text files containg merged smaps from different samples
}
\description{
Mapping Realtionship to unique nanoIDs
}
\examples{
FamilyInfoPrep(
Samplecodes = system.file("extdata", "nanotatoR_sample_codes.csv", package="nanotatoR"),
mergeKey = system.file("extdata", "nanotatoR_control_sample_codes.csv", package="nanotatoR"),
outMode = c("dataframe"))
}
