% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/gene_extract.r
=======
% Please edit documentation in R/entrez_extract.r
>>>>>>> 714868ed579a092f4cd6df397baa7f46eded618f
\name{gene_extraction}
\alias{gene_extraction}
\title{Extracting genes from gene database NCBI.}
\usage{
gene_extraction(terms)
}
\arguments{
<<<<<<< HEAD
\item{terms}{Single or Multiple Terms.}
}
\value{
Dataframe returned containing gene lists in entrezid and Gene 
Symbols, and terms associated with it
=======
\item{terms}{character. Single or Multiple Terms.}
}
\value{
Dataframe returned containing gene lists in entrezid and Gene 
        Symbols, and terms associated with it
>>>>>>> 714868ed579a092f4cd6df397baa7f46eded618f
}
\description{
Extracting genes from gene database NCBI.
}
\examples{
<<<<<<< HEAD
terms="Liver cirrhosis"
ge <- gene_extraction(terms)
=======
\dontrun{
terms="Muscle Weakness"
gene_extraction(terms)
}
>>>>>>> 714868ed579a092f4cd6df397baa7f46eded618f
}
