% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_extract.r
\name{gene_list_generation}
\alias{gene_list_generation}
\title{Extracting genes for phenotype/diseases from NCBI.}
\usage{
gene_list_generation(
  method_entrez = c("Single", "Multiple", "Text"),
  termPath,
  omimID = NULL,
  term,
  outpath,
  thresh = 5,
  returnMethod = c("Text", "dataFrame"),
  omim,
  clinvar,
  gtr,
  removeClinvar = FALSE,
  removeGTR = FALSE,
  downloadClinvar = FALSE,
  downloadGTR = FALSE,
  url_gtr
)
}
\arguments{
\item{method_entrez}{character. Input Method for terms. Choices are 
"Single","Multiple" and "Text".}

\item{termPath}{character. Path and file name for textfile.}

\item{omimID}{numeric. mimID for disease. Default is NULL.}

\item{term}{character. Single or Multiple Terms.}

\item{outpath}{character. Path where gene lists are saved.}

\item{thresh}{integer. Threshold for the number of terms sent to entrez.
Note if large lists are sent to ncbi, it might fail to get
               processed. Default is 5.}

\item{returnMethod}{Method of returning output. Options, Text or data.frame.}

\item{omim}{character. omim2gene file name and location.}

\item{clinvar}{character. clinvar file name and location.}

\item{gtr}{character. gtr file name and location.}

\item{removeClinvar}{logical. Deletes the Clinvar database if TRUE.}

\item{removeGTR}{logical. Deletes the GTR database if TRUE.}

\item{downloadClinvar}{logical. Downloads the Clinvar database if TRUE.}

\item{downloadGTR}{logical. Downloads the GTR database if TRUE.}

\item{url_gtr}{character. url for GTR.}
}
\value{
Text files containg gene list and terms associated with them
        are stored as text files.
}
\description{
Extracting genes for phenotype/diseases from NCBI.
}
\examples{
\dontrun{
terms="CIRRHOSIS, FAMILIAL"
genes <- gene_list_generation(
     method_entrez = c("Single"), 
     term = terms,
     returnMethod=c("dataFrame"), 
     omimID = "OMIM:118980",
       omim = system.file("extdata", "mim2gene.txt", package="nanotatoR"), 
     clinvar = system.file("extdata", "localPDB/", package="nanotatoR"), 
       gtr = system.file("extdata", "gtrDatabase.txt", package="nanotatoR"),
       downloadClinvar = FALSE, downloadGTR = FALSE)
}
}
