% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_extract.r
\name{gtr_gene}
\alias{gtr_gene}
\title{Extracting genes from gtr database NCBI.}
\usage{
gtr_gene(terms, gtr, url_gtr, downloadGTR = TRUE)
}
\arguments{
\item{terms}{Single or Multiple Terms.}

\item{gtr}{character gtr database location.}

\item{url_gtr}{character url for gtr database.}

\item{downloadGTR}{boolean If TRUE, download the gtr database.
Default FALSE.}
}
\value{
Dataframe returned containing gene lists in entrezid and Gene 
Symbols, and terms associated with it
}
\description{
Extracting genes from gtr database NCBI.
}
\examples{
terms="Muscle Weakness"
gtr = system.file("extdata", "gtrDatabase.txt", package="nanotatoR")
ge <- gtr_gene(terms = terms,gtr = gtr, downloadGTR = FALSE)
}
