% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_frequency.r
\name{internalFrequency}
\alias{internalFrequency}
\title{Calculates the internal frequencies of SV in internal cohorts}
\usage{
internalFrequency(mergedFiles, smappath, smapName,
  buildSVInternalDB = FALSE, smapdata, input_fmt_INF = c("Text",
  "dataFrame"), soloPath, solopattern, outpath, win_indel = 10000,
  win_inv_trans = 50000, perc_similarity = 0.5, indelconf = 0.5,
  invconf = 0.01, limsize = 1000, transconf = 0.1,
  returnMethod_Internal = c("Text", "dataFrame"))
}
\arguments{
\item{mergedFiles}{character. Path to the merged SV files.}

\item{smappath}{character. path to the query smap file.}

\item{smapName}{character. File name for the smap}

\item{buildSVInternalDB}{boolean. Checking whether the merged solo
file database exist.}

\item{smapdata}{character. dataframe containing smap data, if
input_fmt_INF= dataFrame}

\item{input_fmt_INF}{character. Choice between Text and DataFrame.}

\item{soloPath}{character. Path to the solo file database.}

\item{solopattern}{character. pattern of the file names to merge.}

\item{outpath}{character. Path to merged SV solo datasets.}

\item{win_indel}{Numeric. Insertion and deletion error window.}

\item{win_inv_trans}{Numeric. Inversion and translocation error window.}

\item{perc_similarity}{Numeric . ThresholdPercentage similarity
of the query SV and reference SV.}

\item{indelconf}{Numeric. Threshold for insertion and deletion confidence.}

\item{invconf}{Numeric. Threshold for inversion confidence.}

\item{limsize}{Numeric. Threshold for size limit for the breakpoint,
for checking that the breakpint size is valid or not. Default is 1000 bases.}

\item{transconf}{Numeric. Threshold for translocation confidence.}

\item{returnMethod_Internal}{character. Choice between Text and DataFrame.}
}
\value{
Text file or data frames containing internalFrequency data.
}
\description{
Calculates the internal frequencies of SV in internal cohorts
}
\examples{
path <- system.file("extdata", "SoloFile", package = "nanotatoR")
pattern <- "_hg19.smap"
smapName <- "F1.1_GM24385_DLE-1_P_trio_hg19.smap"
smappath <- system.file("extdata", smapName, package = "nanotatoR")
intFreq <- internalFrequency(smappath = smappath, buildSVInternalDB = TRUE, soloPath = path, smapName = 
smapName, solopattern = pattern, outpath = path, input_fmt_INF = "Text", win_indel = 10000, limsize = 1000, 
win_inv_trans = 50000, perc_similarity = 0.5, indelconf = 0.5, invconf = 0.01, transconf = 0.1,
returnMethod_Internal = "dataFrame")
intFreq[1:2, ]
}
