% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalFrequency_SVMerge.r
\name{internalFrequencyTrio_Duo}
\alias{internalFrequencyTrio_Duo}
\title{Calculates the internal frequencies of SV in internal cohorts,
for SVMerge}
\usage{
internalFrequencyTrio_Duo(
  mergedFiles,
  smappath,
  smap,
  buildSVInternalDB = FALSE,
  smapdata,
  path,
  pattern,
  outpath,
  win_indel = 10000,
  win_inv_trans = 50000,
  perc_similarity = 0.5,
  indelconf = 0.5,
  invconf = 0.01,
  fname,
  limsize = 1000,
  win_indel_parents = 5000,
  win_inv_trans_parents = 40000,
  transconf = 0.1,
  dbOutput = c("dataframe", "text"),
  perc_similarity_parents = 0.9,
  returnMethod = c("Text", "dataFrame"),
  input_fmt_SV = c("Text", "dataFrame"),
  indexfile,
  EnzymeType = c("SVmerge", "SE"),
  labelType = c("SVMerge", "SE", "Both"),
  SVMerge_path,
  SVMerge_pattern,
  SE_path,
  SE_pattern,
  Samplecodes,
  mergeKey,
  mergedKeyoutpath,
  mergedKeyFname
)
}
\arguments{
\item{mergedFiles}{character. Path to the merged SV files.}

\item{smappath}{character. path to the query smap file.}

\item{smap}{character. File name for the smap}

\item{buildSVInternalDB}{boolean. Checking whether the merged solo 
file database exist.}

\item{smapdata}{character. Dataframe if input type chosen as dataframe.}

\item{path}{character. Path to the solo file database.}

\item{pattern}{character. pattern of the file names to merge.}

\item{outpath}{character. Path where the merged samples are kept.}

\item{win_indel}{N umeric. Insertion and deletion error window.
Default 10000.}

\item{win_inv_trans}{Numeric. Inversion and translocation error window.
Default 50000.}

\item{perc_similarity}{Numeric . ThresholdPercentage similarity 
of the query SV and reference SV. Default 0.5.}

\item{indelconf}{Numeric. Threshold for insertion and deletion confidence.
Default 0.5}

\item{invconf}{Numeric. Threshold for inversion confidence.Default 0.01.}

\item{fname}{character. Filename in case dbOutput = Text.}

\item{limsize}{Numeric. Minimum size of SV that can be determined 
acurately by the Bionano SV caller. Default 1000.}

\item{win_indel_parents}{Numeric. Insertion and deletion error window to 
determine zygosity in case of parents. Default 5000.}

\item{win_inv_trans_parents}{Numeric. Inversion and translocation error 
window to determine zygosity in case of parents. Default 40000.}

\item{transconf}{Numeric. Threshold for translocation confidence. 
Default 0.1.}

\item{dbOutput}{character. Output of merged bionano data.}

\item{perc_similarity_parents}{Numeric . ThresholdPercentage similarity 
for zygosity determination. Default 0.9.}

\item{returnMethod}{character. Choice between Text and DataFrame.
Required if you want to calculate internal frequency.}

\item{input_fmt_SV}{Format in which data 
is provided as an input to thefunction.}

\item{indexfile}{File containing connection between sample and nanoIDs}

\item{EnzymeType}{Character. Type of enzyme. Options Dual and DLE.}

\item{labelType}{character. Type of labels used for mapping. 
Choices are Dual, DLE and Both.}

\item{SVMerge_path}{character. Path for the Dual labelled cmap}

\item{SVMerge_pattern}{character. pattern of the dual files.}

\item{SE_path}{character. Path for the Dual labelled cmap}

\item{SE_pattern}{character. pattern of the dual files.}

\item{Samplecodes}{character. File containing relations and IDs 
associated to them.}

\item{mergeKey}{character. File containing sample ID and relation.}

\item{mergedKeyoutpath}{character. File path storing sample name and nanoID
key information.}

\item{mergedKeyFname}{character. File name storing sample name and nanoID
key information.}
}
\value{
Calculated internal frequency in dataframe or text.
}
\description{
Calculates the internal frequencies of SV in internal cohorts,
for SVMerge
}
\examples{
smapName="GM24385_Ason_DLE1_VAP_trio5.smap"
smap = system.file("extdata", smapName, package="nanotatoR")
indelconf = 0.5; invconf = 0.01;transconf = 0.1;input_fmt_SV="Text";
datInf <- internalFrequency_solo( smap = smap, 
buildSVInternalDB = FALSE, win_indel=10000, 
win_inv_trans=50000, EnzymeType = "SE",
mergedFiles = system.file("extdata", "nanotatoRControl.txt", package="nanotatoR"), 
perc_similarity_parents =0.9,
indexfile = system.file("extdata", "Sample_index.csv", package="nanotatoR"),
perc_similarity=0.5, indelconf=0.5, invconf=0.01, 
transconf=0.1, limsize=1000, win_indel_parents=5000,input_fmt="Text",
win_inv_trans_parents=40000,
returnMethod="dataFrame", input_fmt_SV = "Text")
}
