% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalFrequency_Dual.r
\name{internalFrequencyTrio_Dyad}
\alias{internalFrequencyTrio_Dyad}
\title{Calculates the internal frequencies of SV in internal cohorts}
\usage{
internalFrequencyTrio_Dyad(
  mergedFiles,
  smappath,
  smap,
  buildSVInternalDB = FALSE,
  smapdata,
  input_fmt = c("Text", "dataFrame"),
  path,
  pattern,
  win_indel = 10000,
  win_inv_trans = 50000,
  perc_similarity = 0.5,
  indelconf = 0.1,
  invconf = 0.01,
  fname,
  limsize = 1000,
  win_indel_parents = 5000,
  win_inv_trans_parents = 40000,
  transconf = 0.1,
  dbOutput = c("dataframe", "text"),
  returnMethod = c("Text", "dataFrame"),
  perc_similarity_parents = 0.9,
  indexfile = "X:/Hayks_Materials/BNG/Projects/Sample_NIDKeys.csv",
  Dual_path = "X:/Hayks_Materials/BNG/Projects/VAP_SVmerge_solo_TXTs/",
  labelType = c("Dual", "DLE", "Both"),
  mergeKey = "X:/Hayks_Materials/BNG/Projects/MergeKey.csv",
  outpath = "Y:/Hayks_Materials/BNG/Projects/VAP_SVmerge_solo_TXTs/Merged",
  mergedKeyoutpath = "Y:/Hayks_Materials/BNG/Projects/",
  mergedKeyFname = "mergedKeynanotatoR.csv ",
  EnzymeType = c("Dual", "DLE")
)
}
\arguments{
\item{mergedFiles}{character. Path to the merged SV files.}

\item{smappath}{character. path to the query smap file.}

\item{smap}{character. File name for the smap}

\item{buildSVInternalDB}{boolean. Checking whether the merged solo 
file database exist.}

\item{smapdata}{character. Dataframe if input type chosen as dataframe.}

\item{input_fmt}{Format in which data is provided as an input to the 
function.}

\item{path}{character. Path to the solo file database.}

\item{pattern}{character. pattern of the file names to merge.}

\item{win_indel}{Numeric. Insertion and deletion error window.
Default 10000.}

\item{win_inv_trans}{Numeric. Inversion and translocation error window.
Default 50000.}

\item{perc_similarity}{Numeric . ThresholdPercentage similarity 
of the query SV and reference SV. Default 0.5.}

\item{indelconf}{Numeric. Threshold for insertion and deletion confidence.
Default 0.5}

\item{invconf}{Numeric. Threshold for inversion confidence.Default 0.01.}

\item{fname}{character. Filename in case dbOutput = Text.}

\item{limsize}{Numeric. Minimum size of SV that can be determined 
acurately by the Bionano SV caller. Default 1000.}

\item{win_indel_parents}{Numeric. Insertion and deletion error window to 
determine zygosity in case of parents. Default 5000.}

\item{win_inv_trans_parents}{Numeric. Inversion and translocation error 
window to determine zygosity in case of parents. Default 40000.}

\item{transconf}{Numeric. Threshold for translocation confidence. 
Default 0.1.}

\item{dbOutput}{character. Output of merged bionano data.}

\item{returnMethod}{character. Choice between Text and DataFrame.}

\item{perc_similarity_parents}{Numeric . ThresholdPercentage similarity 
for parent zygosity calculation. Default threshold 0.9.}

\item{indexfile}{File containing connection between sample and nanoIDs}

\item{Dual_path}{character. Path for the Dual labelled cmap}

\item{labelType}{character. Type of labels used for mapping. 
Choices are Dual, DLE and Both.}

\item{outpath}{character. Path to merged SV solo datasets.}

\item{mergedKeyoutpath}{character. File path storing sample name and nanoID
key information.}

\item{mergedKeyFname}{character. File name storing sample name and nanoID
key information.}

\item{EnzymeType}{Character. Type of enzyme. Options Dual and DLE.}

\item{Dual_pattern}{character. pattern of the dual files.}

\item{Samplecodes}{character. File containing relations and IDs 
associated to them.}
}
\value{
Text file or data frames containing internalFrequency data.
Required if you want to calculate internal frequency.
}
\description{
Calculates the internal frequencies of SV in internal cohorts
}
\examples{
mergedFiles <- system.file("extdata", "nanotatoR_merged.txt", package = "nanotatoR")
smapName <- "F1.1_TestSample1_solo_hg19.smap"
smappath <- system.file("extdata", package = "nanotatoR")
indelconf = 0.5; invconf = 0.01;transconf = 0.1;input_fmt="Text";
internalFrequency(mergedFiles = mergedFiles, smappath = smappath , smap = smapName, 
buildSVInternalDB=FALSE, win_indel=10000, 
win_inv_trans=50000, 
perc_similarity=0.5, indelconf=0.5, invconf=0.01, 
transconf=0.1, limsize=1000, win_indel_parents=5000,input_fmt="Text",
win_inv_trans_parents=40000,
returnMethod="dataFrame", perc_similarity_parents = 0.9,
mergedKeyoutpath = "Y:/Hayks_Materials/BNG/Projects/", 
mergedKeyFname = "mergedKeynanotatoR.csv ")
}
