% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortDB_comp.r
\name{makeInternalBNDatabase}
\alias{makeInternalBNDatabase}
\title{Merges Solo SV files to one common SV file}
\usage{
makeInternalBNDatabase(
  path,
  pattern,
  outpath,
  fname,
  dbOutput = c("dataframe", "text")
)
}
\arguments{
\item{path}{character. Path to the solo files.}

\item{pattern}{character. file name pattern for solo files.}

\item{outpath}{character. file path for the output file.}

\item{fname}{character. file name for the output file.}

\item{dbOutput}{character. Output option database or text.}
}
\value{
Text file containing all the solo SMAP files.
}
\description{
Merges Solo SV files to one common SV file
}
\examples{
path <- system.file("extdata", "Bionano_config/", package = "nanotatoR")
pattern <- "_hg19.txt"
mergedSmap <- makeInternalBNDatabase(path = path, 
   pattern = pattern, dbOutput = "dataframe")
mergedSmap[1,]
}
