% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortDB_comp_WIP.r
\name{makeInternalDatabase}
\alias{makeInternalDatabase}
\title{Merges Solo SV files to one common SV file}
\usage{
makeInternalDatabase(
  path,
  pattern,
  outpath,
  fname,
  dbOutput = c("dataframe", "text")
)
}
\arguments{
\item{path}{character. Path to the solo files.}

\item{pattern}{character. file name pattern for solo files.}
}
\value{
Text file containing all the solo SMAP files.
}
\description{
Merges Solo SV files to one common SV file
}
\examples{
path="Z:/bionano/BNG_tools/variant_annotation/08222018/config/"
pattern="_hg19_anonymize.txt"
outpath="Z:/bionano/BNG_tools/variant_annotation/08222018/config/"
makeMergedSmapData(path, pattern, outpath=systemDir())
}
