% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortDB_comp.r
\name{makeMergedSmapData}
\alias{makeMergedSmapData}
\title{Merges Bionano SV files to one common SV file}
\usage{
makeMergedSmapData(path, pattern, outpath, fname,
  dbOutput = c("dataframe", "text"))
}
\arguments{
\item{path}{character. Path to the solo files.}

\item{pattern}{character. file name pattern for solo files.}

\item{outpath}{character. path for output file if dbOutput = text.}

\item{fname}{character. filename if dbOutput = text.}

\item{dbOutput}{character. Output type. Default text and dataframe.}
}
\value{
Text file containing all the solo SMAP files.
}
\description{
Merges Bionano SV files to one common SV file
}
\examples{
path <- system.file("extdata", "Bionano_config/", package = "nanotatoR")
pattern <- "_hg19.txt"
mergedSmap <- makeMergedSmapData(path, pattern, outpath=path,
dbOutput = "dataframe")
}
