% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_frequency.r
\name{makeMergedSVData}
\alias{makeMergedSVData}
\title{Merges Solo SV files to one common SV file}
\usage{
makeMergedSVData(path, pattern, outpath)
}
\arguments{
\item{path}{character. Path to the solo files.}

\item{pattern}{character. file name pattern for solo files.}
}
\value{
Text file containing all the solo SMAP files.
}
\description{
Merges Solo SV files to one common SV file
}
\examples{
\dontrun{
path="Z:/Hayks_Materials/Bionano/Projects/UDN/VSVM_Solo"
pattern="_solo.txt"
makeMergedSVData(path, pattern, outpath=systemDir())
}
}
