% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergesmap.R
\name{Merging_DLE_Dual}
\alias{Merging_DLE_Dual}
\title{Merging Dual and DLE, and adding nanotatoR relation ID}
\usage{
Merging_DLE_Dual(
  labelType = c("Dual", "DLE", "Both"),
  Dual_path = "Y:/Hayks_Materials/BNG/Projects/Control/Solo",
  Dual_pattern = "*.txt",
  DLE_path = "Y:/Hayks_Materials/BNG/Projects/Control/Solo",
  DLE_pattern = "*_DLE_*",
  Samplecodes = "Y:/Hayks_Materials/BNG/Projects/nanotatoR_sample_codes.csv",
  mergeKey = "Y:/Hayks_Materials/BNG/Projects/mergeKey_nanotatoR_Control.csv",
  outpath = "Y:/Hayks_Materials/BNG/Projects/VAP_solos/Merged",
  mergedKeyoutpath = "Y:/Hayks_Materials/BNG/Projects/",
  mergedKeyFname = "nanotatoR_control_sample_codes.csv",
  filename = "nanotatoRpaper_merged_09232019.txt"
)
}
\arguments{
\item{labelType}{character. Type of labels used for mapping. 
Choices are Dual, DLE and Both.}

\item{Dual_path}{character. Path for the Dual labelled cmap}

\item{Dual_pattern}{character. pattern of the dual files.}

\item{Samplecodes}{character. File containing relations and IDs 
associated to them.}

\item{mergeKey}{character. File containing sample ID and relation.}

\item{outpath}{character. Path where the merged samples are kept.}

\item{mergedKeyoutpath}{character. File path storing sample name and nanoID
key information.}

\item{mergedKeyFname}{character. File name storing sample name and nanoID
key information.}

\item{filename}{character. Output file name.}
}
\value{
Text files containg merged smaps from different samples
}
\description{
Merging Dual and DLE, and adding nanotatoR relation ID
}
\examples{
Merging_DLE_Dual (
labelType = c("Both"),
Dual_path = "Y:/Hayks_Materials/BNG/Projects/VAP_SVmerge_solo_TXTs/",
Dual_pattern = "*.txt", 
DLE_path = "Y:/Hayks_Materials/BNG/Projects/VAP_DLE1_solo_SMAPs/",
DLE_pattern = "*.smap",
Samplecodes = "Y:/Hayks_Materials/BNG/Projects/nanotatoR_sample_codes.csv",
mergeKey = "Y:/Hayks_Materials/BNG/Projects/mergedKey_09122019.csv",
outpath = "Y:/Hayks_Materials/BNG/Projects/VAP_SVmerge_solo_TXTs/Merged",
mergedKeyoutpath = "Y:/Hayks_Materials/BNG/Projects/", 
mergedKeyFname = "Sample_NIDKeys.csv",
filename= "UDN_DSD_merged_06232019.txt")
}
