% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergesmap_SVMerge_SE.R
\name{merging_SE_SVMerge}
\alias{merging_SE_SVMerge}
\title{Merging Dual and DLE, and adding nanotatoR relation ID}
\usage{
merging_SE_SVMerge(
  labelType = c("SVMerge", "SE", "Both"),
  SVMerge_path,
  SVMerge_pattern,
  SE_path,
  SE_pattern,
  Samplecodes,
  mergeKey,
  outpath,
  mergedKeyoutpath,
  mergedKeyFname,
  filename,
  outputMode = c("dataframe", "Text")
)
}
\arguments{
\item{labelType}{character. Type of labels used for mapping. 
Choices are Dual, DLE and Both.}

\item{SVMerge_path}{character. Path for the Dual labelled cmap}

\item{SVMerge_pattern}{character. pattern of the dual files.}

\item{SE_path}{character. Path for the Dual labelled cmap}

\item{SE_pattern}{character. pattern of the dual files.}

\item{Samplecodes}{character. File containing relations and IDs 
associated to them.}

\item{mergeKey}{character. File containing sample ID and relation.}

\item{outpath}{character. Path where the merged samples are kept.}

\item{mergedKeyoutpath}{character. File path storing sample name and nanoID
key information.}

\item{mergedKeyFname}{character. File name storing sample name and nanoID
key information.}

\item{filename}{character. Output file name.}

\item{outputMode}{character. Mode of databse output. Text or dataframe.}
}
\value{
Text files containg merged smaps from different samples
}
\description{
Merging Dual and DLE, and adding nanotatoR relation ID
}
\examples{
dat1 <- merging_SE_SVMerge (
labelType = c("SE"),
SE_path = system.file("extdata", "SoloFile/", package="nanotatoR"),
SE_pattern = "*_DLE1_*",
Samplecodes = system.file("extdata", "nanotatoR_sample_codes.csv", package="nanotatoR"),
mergeKey = system.file("extdata", "nanotatoR_control_sample_codes.csv", package="nanotatoR"),
outpath = system.file("extdata", package="nanotatoR"),
mergedKeyoutpath = system.file("extdata", package="nanotatoR"), 
mergedKeyFname = "Sample_index.csv",
filename= "nanotatoRControl.txt", 
outputMode = "dataframe")
}
