% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergesmap.R
\name{mergingSMAP_DLE}
\alias{mergingSMAP_DLE}
\title{Merging DLE labelled smaps}
\usage{
mergingSMAP_DLE(
  path = "Y:/Hayks_Materials/BNG/Projects/VAP_DLE1_solo_SMAPs/",
  pattern = "*_DLE1_*",
  outMode = c("Text", "dataframe"),
  outpath = "Y:/Hayks_Materials/BNG/Projects/VAP_DLE1_solo_SMAPs/Merged"
)
}
\arguments{
\item{path}{character. Path to the solo files directory.}

\item{pattern}{character. Pattern for the solo files.}

\item{outMode}{character. The ouput mode. Choices, dataframe or Text.}

\item{outpath}{character. Path where the dual labelled 
merged samples are kept. Is mandatory if outMode is Text.}
}
\value{
Text files containg merged smaps from different samples
}
\description{
Merging DLE labelled smaps
}
\examples{
mergedSmap <- mergingSMAP_DLE (
path = "Y:/Hayks_Materials/BNG/Projects/VAP_DLE1_solo_SMAPs/",
pattern = "*.txt", outMode = "dataframe", 
outpath = "Y:/Hayks_Materials/BNG/Projects/VAP_SVmerge_solo_TXTs/Merged")
}
