% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergesmap.R
\name{mergingSMAP_Dual}
\alias{mergingSMAP_Dual}
\title{Merging dual labelled smaps}
\usage{
mergingSMAP_Dual(
  path = "Y:/Hayks_Materials/BNG/Projects/VAP_SVmerge_solo_TXTs/",
  pattern = "*.txt",
  outMode = c("dataframe", "Text"),
  outpath = "Y:/Hayks_Materials/BNG/Projects/VAP_SVmerge_solo_TXTs/Merged"
)
}
\arguments{
\item{path}{character. Path to the solo files directory.}

\item{pattern}{character. Patternna for the solo files.}

\item{outMode}{character. The ouput mode. Choices, dataframe or Text.}

\item{outpath}{character. Path where the dual labelled merged samples are kept.
Is mandatory if outMode is Text.}
}
\value{
Text files containg merged smaps from different samples
}
\description{
Merging dual labelled smaps
}
\examples{
mergingSMAP_Dual(
path = "Y:/Hayks_Materials/BNG/Projects/VAP_SVmerge_solo_TXTs/",
pattern = "*.txt", outMode = c("dataframe","Text"), 
outpath = "Y:/Hayks_Materials/BNG/Projects/VAP_SVmerge_solo_TXTs/Merged")
}
