% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessInternalDB.r
\name{mergingSMAP_internal}
\alias{mergingSMAP_internal}
\title{Merging DLE labelled smaps}
\usage{
mergingSMAP_internal(
  path,
  pattern,
  outMode = c("Text", "dataframe"),
  outpath,
  pipeline = c("DVP", "RVP")
)
}
\arguments{
\item{path}{character. Path to the solo files directory.}

\item{pattern}{character. Pattern for the solo files.}

\item{outMode}{character. The ouput mode. Choices, dataframe or Text.}

\item{outpath}{character. Path where the dual labelled 
merged samples are kept. Is mandatory if outMode is Text.}

\item{pipeline}{Analysis pieline used. Options Rare variant Pipeline (RVP)
or Denovo Variant pipeline(DVP).}
}
\value{
Text files containg merged smaps from different samples
}
\description{
Merging DLE labelled smaps
}
\examples{
mergedSmap <- mergingSMAP_SE (
path = system.file("extdata", "SoloFile/", package="nanotatoR"),
pattern = "*_DLE1_*", outMode = "dataframe", 
outpath = system.file("extdata", "Merged/", package="nanotatoR"))
}
