% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergesmap_SVMerge_SE.R
\name{mergingSMAP_SVMerge}
\alias{mergingSMAP_SVMerge}
\title{Merging dual labelled smaps}
\usage{
mergingSMAP_SVMerge(path, pattern, outMode = c("dataframe", "Text"), outpath)
}
\arguments{
\item{path}{character. Path to the solo files directory.}

\item{pattern}{character. Patternna for the solo files.}

\item{outMode}{character. The ouput mode. Choices, dataframe or Text.}

\item{outpath}{character. Path where the dual labelled merged samples are kept.
Is mandatory if outMode is Text.}
}
\value{
Text files containg merged smaps from different samples
}
\description{
Merging dual labelled smaps
}
\examples{
a <- mergingSMAP_SVMerge(
path = system.file("extdata", "SoloFile/", package="nanotatoR"),
pattern = "*.txt", outMode = "dataframe", 
outpath = system.file("extdata", "SoloFile/", package="nanotatoR"))
}
