% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanotatoR.r
\docType{package}
\name{nanotatoR}
\alias{nanotatoR}
\title{nanotatoR: Annotation package for Bionano Data}
\description{
Annotation of Bionano data using available databases
}
\examples{
path <- system.file("extdata", "Bionano_config/", package = "nanotatoR")
pattern <- "_hg19.txt"
mergedSmap <- makeInternalBNDatabase(path = path, 
   pattern = pattern, dbOutput = "dataframe")
mergedSmap[1,]
}
