% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterread.r, R/filterread_DLE.r,
%   R/filterread_DLE_duo.r, R/filterread_DLE_solo.r, R/filterread_Duo.r,
%   R/filterread_solo_Dual_enzyme.r
\name{nonOverlappingDNGenes}
\alias{nonOverlappingDNGenes}
\title{Extracting terms for genes that overlap SVs}
\usage{
nonOverlappingDNGenes(rr, dngene)

nonOverlappingDNGenes(rr, dngene)

nonOverlappingDNGenes(rr, dngene)

nonOverlappingDNGenes(rr, dngene)

nonOverlappingDNGenes(rr, dngene)

nonOverlappingDNGenes(rr, dngene)
}
\arguments{
\item{rr}{character. dataframe with primary genes and terms
associated.}

\item{ogene}{character. genes that overlap the SV.}
}
\value{
Dataframe with overlapping genes and terms.

Dataframe with overlapping genes and terms.

Dataframe with overlapping genes and terms.

Dataframe with overlapping genes and terms.

Dataframe with overlapping genes and terms.

Dataframe with overlapping genes and terms.
}
\description{
Extracting terms for genes that overlap SVs

Extracting terms for genes that overlap SVs

Extracting terms for genes that overlap SVs

Extracting terms for genes that overlap SVs

Extracting terms for genes that overlap SVs

Extracting terms for genes that overlap SVs
}
\examples{
\dontrun{
terms="Muscle Weakness"
dat_geneList <- gene_list_generation(
     method_entrez = c("Single"), 
     term = terms,
     returnMethod=c("dataFrame"), 
	omim = "Y:/Suro/nanotatoRDatabases/mim2gene.txt", 
	clinvar = "Y:/Suro/nanotatoRDatabases/clinvar_07172019.txt",
rr <- dat_geneList
smapName="F1.1_GM24385_DLE-1_P_trio_hg19.smap"
smap = system.file("extdata", smapName, package="nanotatoR")
bedFile <- system.file("extdata", "Homo_sapiens.Hg19.bed", package="nanotatoR")
outpath <- system.file("extdata",  package="nanotatoR")
datcomp <- compSmapbed(smap, bed=bedFile, inputfmtBed =  "BED", outpath,
n = 3, returnMethod_bedcomp = c("dataFrame"))
dngene <- as.character(r$Downstream_nonOverlapGenes_dist_kb)
dataPGDN <- nonOverlappingDNGenes (rr, dngene)
}
\dontrun{
terms="Muscle Weakness"
dat_geneList <- gene_list_generation(
     method_entrez = c("Single"), 
     term = terms,
     returnMethod=c("dataFrame"), 
	omim = "Y:/Suro/nanotatoRDatabases/mim2gene.txt", 
	clinvar = "Y:/Suro/nanotatoRDatabases/clinvar_07172019.txt",
rr <- dat_geneList
smapName="F1.1_GM24385_DLE-1_P_trio_hg19.smap"
smap = system.file("extdata", smapName, package="nanotatoR")
bedFile <- system.file("extdata", "Homo_sapiens.Hg19.bed", package="nanotatoR")
outpath <- system.file("extdata",  package="nanotatoR")
datcomp <- compSmapbed(smap, bed=bedFile, inputfmtBed =  "BED", outpath,
n = 3, returnMethod_bedcomp = c("dataFrame"))
dngene <- as.character(r$Downstream_nonOverlapGenes_dist_kb)
dataPGDN <- nonOverlappingDNGenes (rr, dngene)
}
\dontrun{
terms="Muscle Weakness"
dat_geneList <- gene_list_generation(
     method_entrez = c("Single"), 
     term = terms,
     returnMethod=c("dataFrame"), 
	omim = "Y:/Suro/nanotatoRDatabases/mim2gene.txt", 
	clinvar = "Y:/Suro/nanotatoRDatabases/clinvar_07172019.txt",
rr <- dat_geneList
smapName="F1.1_GM24385_DLE-1_P_trio_hg19.smap"
smap = system.file("extdata", smapName, package="nanotatoR")
bedFile <- system.file("extdata", "Homo_sapiens.Hg19.bed", package="nanotatoR")
outpath <- system.file("extdata",  package="nanotatoR")
datcomp <- compSmapbed(smap, bed=bedFile, inputfmtBed =  "BED", outpath,
n = 3, returnMethod_bedcomp = c("dataFrame"))
dngene <- as.character(r$Downstream_nonOverlapGenes_dist_kb)
dataPGDN <- nonOverlappingDNGenes (rr, dngene)
}
\dontrun{
terms="Muscle Weakness"
dat_geneList <- gene_list_generation(
     method_entrez = c("Single"), 
     term = terms,
     returnMethod=c("dataFrame"), 
	omim = "Y:/Suro/nanotatoRDatabases/mim2gene.txt", 
	clinvar = "Y:/Suro/nanotatoRDatabases/clinvar_07172019.txt",
rr <- dat_geneList
smapName="F1.1_GM24385_DLE-1_P_trio_hg19.smap"
smap = system.file("extdata", smapName, package="nanotatoR")
bedFile <- system.file("extdata", "Homo_sapiens.Hg19.bed", package="nanotatoR")
outpath <- system.file("extdata",  package="nanotatoR")
datcomp <- compSmapbed(smap, bed=bedFile, inputfmtBed =  "BED", outpath,
n = 3, returnMethod_bedcomp = c("dataFrame"))
dngene <- as.character(r$Downstream_nonOverlapGenes_dist_kb)
dataPGDN <- nonOverlappingDNGenes (rr, dngene)
}
\dontrun{
terms="Muscle Weakness"
dat_geneList <- gene_list_generation(
     method_entrez = c("Single"), 
     term = terms,
     returnMethod=c("dataFrame"), 
	omim = "Y:/Suro/nanotatoRDatabases/mim2gene.txt", 
	clinvar = "Y:/Suro/nanotatoRDatabases/clinvar_07172019.txt",
rr <- dat_geneList
smapName="F1.1_GM24385_DLE-1_P_trio_hg19.smap"
smap = system.file("extdata", smapName, package="nanotatoR")
bedFile <- system.file("extdata", "Homo_sapiens.Hg19.bed", package="nanotatoR")
outpath <- system.file("extdata",  package="nanotatoR")
datcomp <- compSmapbed(smap, bed=bedFile, inputfmtBed =  "BED", outpath,
n = 3, returnMethod_bedcomp = c("dataFrame"))
dngene <- as.character(r$Downstream_nonOverlapGenes_dist_kb)
dataPGDN <- nonOverlappingDNGenes (rr, dngene)
}
\dontrun{
terms="Muscle Weakness"
dat_geneList <- gene_list_generation(
     method_entrez = c("Single"), 
     term = terms,
     returnMethod=c("dataFrame"), 
	omim = "Y:/Suro/nanotatoRDatabases/mim2gene.txt", 
	clinvar = "Y:/Suro/nanotatoRDatabases/clinvar_07172019.txt",
rr <- dat_geneList
smapName="F1.1_GM24385_DLE-1_P_trio_hg19.smap"
smap = system.file("extdata", smapName, package="nanotatoR")
bedFile <- system.file("extdata", "Homo_sapiens.Hg19.bed", package="nanotatoR")
outpath <- system.file("extdata",  package="nanotatoR")
datcomp <- compSmapbed(smap, bed=bedFile, inputfmtBed =  "BED", outpath,
n = 3, returnMethod_bedcomp = c("dataFrame"))
dngene <- as.character(r$Downstream_nonOverlapGenes_dist_kb)
dataPGDN <- nonOverlappingDNGenes (rr, dngene)
}
}
