% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrez_extract.r
\name{omim_gene}
\alias{omim_gene}
\title{Extracting genes from OMIM database NCBI.}
\usage{
omim_gene(terms)
}
\arguments{
\item{terms}{Single or Multiple Terms.}
}
\value{
Dataframe returned containing gene lists in entrezid and Gene 
Symbols, and terms associated with it
}
\description{
Extracting genes from OMIM database NCBI.
}
\examples{
terms="Muscle Weakness"
omim_gene(terms)
}
