% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/OverlappingGenes.r
=======
% Please edit documentation in R/Bed_SV_Comp.r
>>>>>>> 714868ed579a092f4cd6df397baa7f46eded618f
\name{overlapGenes}
\alias{overlapGenes}
\title{Calculates Genes that overlap the SV region}
\usage{
<<<<<<< HEAD
overlapGenes(
  bed,
  chrom,
  startpos,
  endpos,
  svid,
  chrom2,
  SVTyp,
  bperrorindel = 3000,
  bperrorinvtrans = 50000
)
=======
overlapGenes(bed, chrom, startpos, endpos, svid)
>>>>>>> 714868ed579a092f4cd6df397baa7f46eded618f
}
\arguments{
\item{bed}{Text Bionano Bed file.}

\item{chrom}{character SVmap chromosome.}

\item{startpos}{numeric starting position of the breakpoints.}

\item{endpos}{numeric end position of the breakpoints.}

\item{svid}{numeric Structural variant identifier (Bionano generated).}
<<<<<<< HEAD

\item{chrom2}{character SVmap chromosome number 2.}

\item{SVTyp}{Character. Type of SV.}

\item{bperrorindel}{Numeric. base pair error indel.}

\item{bperrorinvtrans}{Numeric. base pair error invtranslocation.}
}
\value{
Data Frame. Contains the SVID,Gene name,strand information and
=======
}
\value{
Data Frame. Contains the SVID,Gene name,strand information and 
>>>>>>> 714868ed579a092f4cd6df397baa7f46eded618f
percentage of SV covered.
}
\description{
Calculates Genes that overlap the SV region
}
\examples{
<<<<<<< HEAD
smapName="GM24385_Ason_DLE1_VAP_trio5.smap"
smap = system.file("extdata", smapName, package="nanotatoR")
bedFile <- system.file("extdata", "HomoSapienGRCH19_lift37.bed", 
       package="nanotatoR")
bed<-buildrunBNBedFiles(bedFile,returnMethod="dataFrame")
smap<-readSMap_DLE(smap, input_fmt_smap = "Text")
chrom<-smap$RefcontigID1
chrom2 <- smap$RefcontigID2
startpos<-smap$RefStartPos
endpos<-smap$RefEndPos
if (length(grep("SVIndex",names(smap)))>0){
       svid <- smap$SVIndex
   }else{
    svid <- smap$SmapEntryID
    }
SVTyp <- smap$Type
overlapGenes(bed = bed, chrom = chrom, startpos = startpos, 
   endpos = endpos, chrom2 = chrom2, svid = svid,
   SVTyp = SVTyp, 
   bperrorindel = 3000, bperrorinvtrans = 50000)
=======
\dontrun{
smap="C:\\\\Annotator\\\\Data\\\\F1.1_UDN287643_P_Q.S_VAP_SVmerge_trio_original.txt"
bedFile = "C:\\\\Annotator\\\\Data\\\\Homo_sapiens.GRCH19.bed"
bed<-buildrunBNBedFiles(bedFile,returnMethod = "dataFrame")
smap<-readSMap(smap)
chrom<-smap$RefcontigID1
startpos<-smap$RefStartPos
endpos<-smap$RefEndPos
svid<-smap$SVIndex
overlapGenes(bed, chrom, startpos, endpos, svid)
}
>>>>>>> 714868ed579a092f4cd6df397baa7f46eded618f
}
