% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_overlapping_non_overlapping.r
\name{overlappingGenes}
\alias{overlappingGenes}
\title{Extracting terms for genes that overlap SVs}
\usage{
overlappingGenes(rr, ogene)
}
\arguments{
\item{rr}{character. dataframe with primary genes and terms
associated.}

\item{ogene}{character. genes that overlap the SV.}
}
\value{
Dataframe with overlapping genes and terms.
}
\description{
Extracting terms for genes that overlap SVs
}
\examples{
terms= c("steroid_Gene","steroid synthesis_Gene")
genes <- c("NR1H3", "ABCC4")
rr <- data.frame(Genes = genes, Terms = terms)
smapName="GM24385_Ason_DLE1_VAP_trio5.smap"
smap = system.file("extdata", smapName, package="nanotatoR")
bedFile <- system.file("extdata", "HomoSapienGRCH19_lift37.bed", package="nanotatoR")
outpath <- system.file("extdata", package="nanotatoR")
datcomp<-overlapnearestgeneSearch(smap = smap, 
bed=bedFile, inputfmtBed = "bed", outpath, 
n = 3, returnMethod_bedcomp = c("dataFrame"), 
input_fmt_SV = "Text",
EnzymeType = "SE", 
bperrorindel = 3000, bperrorinvtrans = 50000)
ogene <- as.character(datcomp$OverlapGenes_strand_perc)
datogenes <- overlappingGenes (rr, ogene)
}
