% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASEQ_Analysis.r
\name{OverlapRNAseq}
\alias{OverlapRNAseq}
\title{Extract Read counts for genes that overlap SVs.}
\usage{
OverlapRNAseq(
  gnsOverlap,
  SVID,
  RNASeqData,
  pattern_Proband = NA,
  pattern_Mother = NA,
  pattern_Father = NA,
  pattern_Sibling = NA
)
}
\arguments{
\item{gnsOverlap}{character. genes that overlap SV.}

\item{SVID}{character. ID of the SVs.}

\item{RNASeqData}{character. Expression of the genes.}

\item{pattern_Proband}{character. Pattern to identify the proband reads.}

\item{pattern_Mother}{character. Pattern to identify the mother reads.}

\item{pattern_Father}{character. Pattern to identify the father reads.}

\item{pattern_Sibling}{character. Pattern to identify the sibling reads.}
}
\value{
Text or Dataframe containing TPM read counts of genes in the family.
}
\description{
Extract Read counts for genes that overlap SVs.
}
\examples{
RNASeqDir="Z:/Suro/Annotator/Data/RNAseq"
returnMethod="dataFrame"
OverlapRNAseq(RNASeqDir,returnMethod)
}
