% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASEQ_Analysis_solo_SVMerge_SE.r
\name{OverlapRNAseq_solo}
\alias{OverlapRNAseq_solo}
\title{Annotating the Overlapping genes with RNAseq expression}
\usage{
OverlapRNAseq_solo(gnsOverlap, SVID, RNASeqData, pattern_Proband = NA)
}
\arguments{
\item{gnsOverlap}{character. Vector containing overlapping genes.}

\item{SVID}{character. SV Index ID.}

\item{RNASeqData}{dataFrame. RNAseq data with gene names.}

\item{pattern_Proband}{character. Pattern for proband.}
}
\value{
Dataframe containing TPM read counts of overlapping genes.
}
\description{
Annotating the Overlapping genes with RNAseq expression
}
\examples{
RNASeqDir = system.file("extdata", package="nanotatoR")
returnMethod="dataFrame"
datRNASeq <- RNAseqcombine_solo(RNASeqDir = RNASeqDir,
returnMethod = returnMethod)
gnsOverlap <- c("AGL")
SVID = 397
datgnovrlap <- OverlapRNAseq_solo(gnsOverlap = gnsOverlap, 
SVID = SVID, RNASeqData = datRNASeq,
pattern_Proband = "*_P_*")
}
