% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoextract_mod.r
\name{phenoextractHPO_mod}
\alias{phenoextractHPO_mod}
\title{Extract the genes related to a disease or disease alias from HPO database.}
\usage{
phenoextractHPO_mod(keyword, localPDB.path)
}
\arguments{
\item{keyword}{character. character string: keyword,
to search a disease, a clinical feature, or a phenotype.}

\item{localPDB.path}{character. the path of 
localized public data bases. The default value is set in the working 
directory.}
}
\value{
subset of HPO and 
extract the genes and alias for a disease(phenotype), 
or a clinical feature. Function modified from 
pheno_extract_HPO function VarFromPDB.
}
\description{
Extract the genes related to a disease or disease alias from HPO database.
}
\examples{
HPO.phenotype = phenoextractHPO_mod("retinoblastoma",
localPDB.path = system.file("extdata", "localPDB", package="nanotatoR"))
}
