% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/Bed_SV_Comp_WIP.r
=======
% Please edit documentation in R/Bed_SV_Comp.r
>>>>>>> 04c1b43bc3aded9fbd47aace30093a5b16389a06
\name{readBNBedFiles}
\alias{readBNBedFiles}
\title{Reads Bionano Bedfiles}
\usage{
readBNBedFiles(BNFile)
}
\arguments{
\item{BNFile}{character. Path to Bionano Bed File.}
}
\value{
Data Frame Contains the gene information.
}
\description{
Reads Bionano Bedfiles
}
\examples{
<<<<<<< HEAD
BNFile <- system.file("extdata", "Homo_sapiens.GRCH19_BN.bed", package="nanotatoR")
=======
BNFile <- system.file("extdata", "Homo_sapiens.Hg19_BN_bed.txt", 
package="nanotatoR")
>>>>>>> 04c1b43bc3aded9fbd47aace30093a5b16389a06
bed<-readBNBedFiles(BNFile)
}
