% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrez_extract.r
\name{reading_GTR}
\alias{reading_GTR}
\title{Reading and parsing gtr database.}
\usage{
reading_GTR(
  gtr,
  downloadGTR,
  url_gtr = "ftp://ftp.ncbi.nlm.nih.gov/pub/GTR/data/test_condition_gene.txt"
)
}
\arguments{
\item{gtr}{character gtr database location.}

\item{downloadGTR}{logical if true, downloads gtr database, .
and store data in the gtr location, else reads dataset from
gtr location.}
}
\value{
Dataframe representation of gtr database.
}
\description{
Reading and parsing gtr database.
}
\examples{
a <- reading_GTR(gtr  = "Y:/Suro/nanotatoRDatabases/gtr_07162019.txt",
       downloadGTR = FALSE)
}
