% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrez_extract.r
\name{reading_mim2gene}
\alias{reading_mim2gene}
\title{Reading and parsing OMIM database.}
\usage{
reading_mim2gene(omim)
}
\arguments{
\item{omim}{character omim database location.}
}
\value{
Dataframe returned containing Omim ID and gene IDs.
}
\description{
Reading and parsing OMIM database.
}
\examples{
reading_mim2gene(omim  = "Y:/Suro/nanotatoRDatabases/mim2gene.txt")
}
