% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bed_SV_Comp.r
\name{readSMap}
\alias{readSMap}
\title{Reads SMAP files to extract information}
\usage{
readSMap(smap)
}
\arguments{
\item{smap}{character. Path to SMAP file.}
}
\value{
Data Frame or text file. Contains the SMAP information.
}
\description{
Reads SMAP files to extract information
}
\examples{
\dontrun{
smap = "C:\\\\Annotator\\\\Data\\\\F1_UDN287643_P_Q.S_VAP_SVmerge_trio.txt"
readSMap(smap)
}
}
