% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OverlappingGenes.r
\name{readSMap_DLE}
\alias{readSMap_DLE}
\title{Reads DLE SMAP files to extract information}
\usage{
readSMap_DLE(smap, smapdata, input_fmt_smap = "Text")
}
\arguments{
\item{smap}{character. Path to SMAP file.}

\item{smapdata}{dataframe. variable for smap dataset.}

\item{input_fmt_smap}{character. input format for smap
text or dataframe.}
}
\value{
Data Frame or text file. Contains the SMAP information.
}
\description{
Reads DLE SMAP files to extract information
}
\examples{
smapName="GM24385_Ason_DLE1_VAP_trio5.smap"
smap = system.file("extdata", smapName, package="nanotatoR")
readSMap_DLE(smap, input_fmt_smap = "Text")
}
