% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASEQ_Analysis.r
\name{RNAseqcombine}
\alias{RNAseqcombine}
\title{Combining the RNAseq reads of family members in a 
single file.}
\usage{
RNAseqcombine(
  RNASeqDir,
  returnMethod = c("Text", "dataFrame"),
  outpath = "",
  outFileName = ""
)
}
\arguments{
\item{RNASeqDir}{character. Directory containing RNAseq reads.}

\item{returnMethod}{character. Method of returning Data.}

\item{outpath}{character. Contains file path if Method of return is chosen as 
Text.}

\item{outFileName}{character. Output file name.}
}
\value{
Text or Dataframe containing TPM read counts of genes in the family.
}
\description{
Combining the RNAseq reads of family members in a 
single file.
}
\examples{
RNASeqDir="Z:/Suro/Annotator/Data/RNAseq"
returnMethod="dataFrame"
RNAseqcombine(RNASeqDir,returnMethod)
}
