% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASEQ_Analysis_solo_SVMerge_SE.r
\name{RNAseqcombine_solo}
\alias{RNAseqcombine_solo}
\title{Combining the RNAseq reads of family members in a 
single file.}
\usage{
RNAseqcombine_solo(
  RNASeqDir,
  returnMethod = c("Text", "dataFrame"),
  outpath = "",
  outFileName = ""
)
}
\arguments{
\item{RNASeqDir}{character. Directory containing RNAseq reads.}

\item{returnMethod}{character. Method of returning Data.}

\item{outpath}{character. Contains file path if Method of return is chosen as 
Text.}

\item{outFileName}{character. Output file name.}
}
\value{
Text or Dataframe containing TPM read counts of genes in the family.
}
\description{
Combining the RNAseq reads of family members in a 
single file.
}
\examples{
RNASeqDir = system.file("extdata", package="nanotatoR")
returnMethod="dataFrame"
datRNASeq <- RNAseqcombine_solo(RNASeqDir = RNASeqDir,
returnMethod = returnMethod)
}
