% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterread.r
\name{run_bionano_filter}
\alias{run_bionano_filter}
\title{Getting the data from annotated smaps to extract SV
information based on type of variants.}
\usage{
run_bionano_filter(input_fmt_geneList = c("Text", "dataFrame"),
  input_fmt_svMap = c("Text", "dataFrame"), SVFile = NULL, svData,
  dat_geneList, fileName, outpath, outputFilename = "", RZIPpath)
}
\arguments{
\item{input_fmt_geneList}{character. Choice of gene list input
Text or Dataframe.}

\item{input_fmt_svMap}{character. Choice of gene list input
Text or Dataframe.}

\item{SVFile}{character. SV file name.}

\item{svData}{Dataframe Input data containing SV data.}

\item{dat_geneList}{Dataframe Input data containing geneList data.}

\item{fileName}{Character Name of file containing Gene List data.}

\item{outpath}{Character Directory to the output file.}

\item{outputFilename}{Character Output filename.}

\item{RZIPpath}{Character Path for the Rtools Zip package.}
}
\value{
Excel file containing the annotated SV map, tabs divided based on
type of SVs.
}
\description{
Getting the data from annotated smaps to extract SV
information based on type of variants.
}
\examples{
\dontrun{
terms <- "Muscle Weakness"
gene <- gene_list_generation(
  method = "Single", term = terms,
  returnMethod_GeneList = "dataFrame"
)
RzipFile = "zip.exe"
RZIPpath <- system.file("extdata", RzipFile, package = "nanotatoR")
smapName <- "F1.1_TestSample1_solo_hg19.smap"
smappath <- system.file("extdata", smapName, package = "nanotatoR")
smappath1 <- system.file("extdata", package = "nanotatoR")
run_bionano_filter(input_fmt_geneList = c("dataframe"), input_fmt_svMap = c("Text"),
  SVFile = smappath, dat_geneList = gene, outpath = smappath1, outputFilename = "test", 
  RZIPpath = RZIPpath)
}
}
