% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterread_solo_Dual_enzyme.r
\name{run_bionano_filter_singleton_Dual}
\alias{run_bionano_filter_singleton_Dual}
\title{Getting the data from annotated smaps to extract SV
information based on type of variants.}
\usage{
run_bionano_filter_singleton_Dual(
  input_fmt_geneList = c("Text", "dataFrame"),
  input_fmt_svMap = c("Text", "dataFrame"),
  SVFile = NULL,
  svData,
  dat_geneList,
  fileName,
  outpath,
  outputFilename = "",
  RZIPpath,
  primaryGenesPresent = TRUE,
  outputType = c("Excel", "csv"),
  directoryName,
  fileprefix
)
}
\arguments{
\item{input_fmt_geneList}{character. Choice of gene list input
Text or Dataframe.}

\item{input_fmt_svMap}{character. Choice of gene list input
Text or Dataframe.}

\item{SVFile}{character. SV file name.}

\item{svData}{Dataframe Input data containing SV data.}

\item{dat_geneList}{Dataframe Input data containing geneList data.}

\item{fileName}{Character Name of file containing Gene List data.}

\item{outpath}{Character Directory to the output file.}

\item{outputFilename}{Character Output filename.}

\item{RZIPpath}{Character Path for the Rtools Zip package.}

\item{outputType}{Variants in excel tabs or in different csv files.
Options Excel or csv.}

\item{directoryName}{Directory name where individual SV files will be stored.}

\item{fileprefix}{Prefix to use for each of the files in the directory.}
}
\value{
Excel file containing the annotated SV map, tabs divided based on
type of SVs.

individual csv files for each variant type if csv option is chosen.
}
\description{
Getting the data from annotated smaps to extract SV
information based on type of variants.
}
\examples{
\dontrun{
terms <- "Muscle Weakness"
gene <- gene_list_generation(
  method = "Single", term = terms,
  returnMethod_GeneList = "dataFrame"
)
smapName <- "F1.1_GM24385_DLE-1_P_trio_hg19.smap"
smappath <- system.file("extdata", smapName, package = "nanotatoR")
run_bionano_filter_singleton_Dual(
       input_fmt_geneList = c("dataFrame"),
       input_fmt_svMap = c("Text"),
       SVFile = NULL, svData, dat_geneList, fileName, outpath,
       outputFilename = "", RZIPpath, primaryGenesPresent = TRUE,
       outputType = c("Excel"))
}
}
