% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterread_Duo_SVMerge.r
\name{run_bionano_filter_SVMerge_duo}
\alias{run_bionano_filter_SVMerge_duo}
\title{Getting the data from annotated smaps to extract SV
information based on type of variants.}
\usage{
run_bionano_filter_SVMerge_duo(
  input_fmt_geneList = c("Text", "dataFrame"),
  input_fmt_SV = c("Text", "dataFrame"),
  smap = NULL,
  svData,
  dat_geneList,
  fileName,
  outpath,
  outputFilename = "",
  RZIPpath,
  outputType = c("Excel", "csv"),
  primaryGenesPresent = TRUE,
  fileprefix,
  directoryName,
  EnzymeType = c("SVMerge", "SE")
)
}
\arguments{
\item{input_fmt_geneList}{character. Choice of gene list input
Text or Dataframe.}

\item{input_fmt_SV}{character. Choice of gene list input
Text or Dataframe.}

\item{smap}{character. SV file name.}

\item{svData}{Dataframe Input data containing SV data.}

\item{dat_geneList}{Dataframe Input data containing geneList data.}

\item{fileName}{Character Name of file containing Gene List data.}

\item{outpath}{Character Directory to the output file.}

\item{outputFilename}{Character Output filename.}

\item{RZIPpath}{Character Path for the Rtools Zip package.}

\item{outputType}{Character. Variants in excel tabs or in different csv files.
Options Excel or csv.}

\item{primaryGenesPresent}{boolean Checks whether the primary gene List is present or not.}

\item{fileprefix}{Character. fileprefix to use for each of the files in the directory.}

\item{directoryName}{Character. Directory name where individual SV files will be stored.}

\item{EnzymeType}{Character. Enzyme type used. Options SVMerge or SE.}
}
\value{
Excel file containing the annotated SV map, tabs divided based on
type of SVs.
}
\description{
Getting the data from annotated smaps to extract SV
information based on type of variants.
}
\examples{
\dontrun{
smapName <- "GM24385_DLE-1_P_trio_hg19.smap"
outputFilename <- "GM24385_DLE-1_P_trio_hg19_out"
smappath <- system.file("extdata", smapName, package = "nanotatoR")
outpath <- system.file("extdata", smapName, package = "nanotatoR")
RZIPpath <- system.file("extdata", "zip.exe", package = "nanotatoR")
run_bionano_filter_SVMerge_duo (input_fmt_geneList = c("Text"),
input_fmt_SV = c("Text"),
smap = smappath, 
dat_geneList = dat_geneList,
RZIPpath = RZIPpath, EnzymeType = c("SVMerge"),
outputType = c("Excel"),
primaryGenesPresent = FALSE, 
outputFilename = outputFilename,
outpath = outpath)
}
}
