% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASEQ_Analysis.r
\name{SmapRNAseqquery}
\alias{SmapRNAseqquery}
\title{Extract Read counts for genes that are near 
or overalapping SVs.}
\usage{
SmapRNAseqquery(
  input_fmt_SV = c("Text", "dataFrame"),
  smapdata,
  smappath,
  input_fmt_RNASeq = c("Text", "dataFrame"),
  RNASeqData,
  RNASeqPATH,
  outputfmt = c("Text", "datFrame"),
  pattern_Proband = NA,
  pattern_Mother = NA,
  pattern_Father = NA,
  pattern_Sibling = NA,
  EnzymeType = c("Dual", "DLE")
)
}
\arguments{
\item{input_fmt_SV}{character. genes that are upstream 
and/or downstream of SV.}

\item{RNASeqData}{character. Expression of the genes.}

\item{pattern_Proband}{character. Pattern to identify the proband reads.}

\item{pattern_Mother}{character. Pattern to identify the mother reads.}

\item{pattern_Father}{character. Pattern to identify the father reads.}

\item{pattern_Sibling}{character. Pattern to identify the sibling reads.}

\item{EnzymeType}{character. Enzyme used. option "Dual" or "DLE".}

\item{SVID}{character. ID of the SVs.}
}
\value{
Text or Dataframe containing TPM read counts of genes in the family.
}
\description{
Extract Read counts for genes that are near 
or overalapping SVs.
}
\examples{
RNASeqDir="Z:/Suro/Annotator/Data/RNAseq"
returnMethod="dataFrame"
RNAseqcombine(RNASeqDir,returnMethod)
}
