% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASEQ_Analysis_solo.r
\name{SmapRNAseqquery_solo}
\alias{SmapRNAseqquery_solo}
\title{Annotating the Overlapping and Non-Overlapping genes with RNAseq expression}
\usage{
SmapRNAseqquery_solo(
  input_fmt_SV = c("Text", "dataFrame"),
  smapdata,
  smappath,
  input_fmt_RNASeq = c("Text", "dataFrame"),
  RNASeqData,
  RNASeqPATH,
  outputfmt = c("Text", "datFrame"),
  pattern_Proband = NA,
  EnzymeType = c("Dual", "DLE")
)
}
\arguments{
\item{input_fmt_SV}{character. Input format of the SV data.Options 
"Text" or "DataFrame".}

\item{smapdata}{dataframe. SV data dataframe.}

\item{smappath}{character. Path to the SV dataset.}

\item{input_fmt_RNASeq}{character. Input format of
the RNASeq data. Options "Text" or "DataFrame"..}

\item{RNASeqData}{dataFrame. RNAseq data with gene names.}

\item{RNASeqPATH}{character. RNAseq dataset path .}

\item{outputfmt}{character. Output format of
the result. Options "Text" or "DataFrame"..}

\item{pattern_Proband}{character. Pattern for proband.}

\item{EnzymeType}{character. Enzyme used. option "Dual" or "DLE".}
}
\value{
Dataframe Annotated datafreme with RNASeq data.
}
\description{
Annotating the Overlapping and Non-Overlapping genes with RNAseq expression
}
\examples{
RNASeqDir="Z:/Suro/Annotator/Data/RNAseq"
returnMethod="dataFrame"
RNAseqcombine(RNASeqDir,returnMethod)
}
