% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGV_extraction.r
\name{UniqueSample}
\alias{UniqueSample}
\title{Calculating number of samples.}
\usage{
UniqueSample(samp)
}
\arguments{
\item{samp}{character. Unique samples}
}
\description{
Calculating number of samples.
}
\examples{
\dontrun{
UniqueSample(samp)
}
}
