% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/Bed_SV_Comp_WIP.r
=======
% Please edit documentation in R/Bed_SV_Comp.r
>>>>>>> 04c1b43bc3aded9fbd47aace30093a5b16389a06
\name{buildrunBNBedFiles}
\alias{buildrunBNBedFiles}
\title{Reads BED files to produce bionano Bed files}
\usage{
<<<<<<< HEAD
buildrunBNBedFiles(bedFile, returnMethod = c("Text", "dataFrame"), outdir)
=======
buildrunBNBedFiles(bedFile, returnMethod = c("Text", "dataFrame"),
  outdir)
>>>>>>> 04c1b43bc3aded9fbd47aace30093a5b16389a06
}
\arguments{
\item{bedFile}{character. Path to UCSC Bed File.}

\item{returnMethod}{character. Path to output directory.}

\item{outdir}{character. Path to output directory.}
}
\value{
Data Frame or text file. Contains the gene information.
}
\description{
Reads BED files to produce bionano Bed files
}
\examples{
bedFile <- system.file("extdata", "Homo_sapiens.Hg19.bed", 
<<<<<<< HEAD
      package="nanotatoR")
=======
package="nanotatoR")
>>>>>>> 04c1b43bc3aded9fbd47aace30093a5b16389a06
bed<-buildrunBNBedFiles(bedFile,returnMethod="dataFrame")
}
