% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bed_SV_Comp.r
\name{buildrunBNBedFiles}
\alias{buildrunBNBedFiles}
\title{Reads BED files to produce bionano Bed files}
\usage{
buildrunBNBedFiles(bedFile, returnMethod = c("Text", "dataFrame"),
  outdir)
}
\arguments{
\item{bedFile}{character. Path to UCSC Bed File.}

\item{outdir}{character. Path to output directory.}
}
\value{
Data Frame or text file. Contains the gene information.
}
\description{
Reads BED files to produce bionano Bed files
}
\examples{
bedFile = "C:\\\\nanotatoR\\\\Data\\\\Homo_sapiens.Hg19.bed"
buildrunBNBedFiles(bedFile,outdir=tempdir())
}
