% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrez_extract.r
\name{clinvar_gene}
\alias{clinvar_gene}
\title{Extracting genes from gtr database NCBI.}
\usage{
clinvar_gene(terms, clinvar, downloadClinvar)
}
\arguments{
\item{terms}{Single or Multiple Terms.}

\item{gtr}{character gtr database location.}

\item{downloadGTR}{boolean If TRUE, download the gtr database.
Default FALSE.}

\item{clinsig.}{character. clinical significance chosen. Choices are
"Pathogenic", "Pathogenic/Likely pathogenic", "Likely pathogenic",
"Benign", "Likely Benign", "Benign/Likely Benign","VUS", 
"Benign/Likely Benign/VUS", "All".}
}
\value{
Dataframe returned containing gene lists in entrezid and Gene 
Symbols, and terms associated with it
}
\description{
Extracting genes from gtr database NCBI.
}
\examples{
terms="Muscle Weakness"
ge <- gtr_gene(terms,gtr = "Y:/Suro/nanotatoRDatabases/gtr_07162019.txt")
}
