% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/gene_extract.r
=======
% Please edit documentation in R/entrez_extract.r
>>>>>>> 714868ed579a092f4cd6df397baa7f46eded618f
\name{clinvar_gene}
\alias{clinvar_gene}
\title{Extracting genes from clinvar database NCBI.}
\usage{
<<<<<<< HEAD
clinvar_gene(terms, clinvar, downloadClinvar, omimID = NULL)
}
\arguments{
\item{terms}{Single or Multiple Terms.}

\item{clinvar}{character clinvar database location.}

\item{downloadClinvar}{boolean If TRUE, download the gtr database.
Default FALSE.}

\item{omimID}{numeric Omim Id for disease.}
}
\value{
Dataframe returned containing gene lists in entrezid and Gene 
Symbols, and terms associated with it
=======
clinvar_gene(terms)
}
\arguments{
\item{terms}{character. Single or Multiple Terms.}
}
\value{
Dataframe returned containing gene lists in entrezid and Gene 
        Symbols, and terms associated with it
>>>>>>> 714868ed579a092f4cd6df397baa7f46eded618f
}
\description{
Extracting genes from clinvar database NCBI.
}
\examples{
\dontrun{
terms="Muscle Weakness"
<<<<<<< HEAD
clinvar = system.file("extdata", "localPDB/", package="nanotatoR")
downloadClinvar = TRUE
ge <- clinvar_gene(terms = terms, clinvar = clinvar, 
downloadClinvar = downloadClinvar, 
omimID = "OMIM:118980")
=======
gtr_gene(terms, dataset="hsapiens_gene_ensembl")
>>>>>>> 714868ed579a092f4cd6df397baa7f46eded618f
}
}
