% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrez_extract.r
\name{clinvar_gene}
\alias{clinvar_gene}
\title{Extracting genes from clinvar database NCBI.}
\usage{
clinvar_gene(terms)
}
\arguments{
\item{terms}{Single or Multiple Terms.}
}
\value{
Dataframe returned containing gene lists in entrezid and Gene 
Symbols, and terms associated with it
}
\description{
Extracting genes from clinvar database NCBI.
}
\examples{
terms="Muscle Weakness"
clinvar_gene(terms)
}
