% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortDB_comp_WIP.r
\name{cohortFrequency}
\alias{cohortFrequency}
\title{Calculates the internal frequencies of SV in internal cohorts}
\usage{
cohortFrequency(
  internalBNDB,
  smappath,
  smap,
  buildBNInternalDB = FALSE,
  smapdata,
  input_fmt = c("Text", "dataFrame"),
  dbOutput = c("dataframe", "text"),
  InternalDBpath,
  InternalDBpattern,
  fname,
  outpath,
  win_indel = 10000,
  win_inv_trans = 50000,
  perc_similarity = 0.5,
  indelconf = 0.5,
  invconf = 0.01,
  limsize = 1000,
  transconf = 0.1,
  returnMethod = c("Text", "dataFrame"),
  EnzymeType = c("Dual", "DLE")
)
}
\arguments{
\item{smappath}{character. path to the query smap file.}

\item{smap}{character. File name for the smap}

\item{outpath}{character. Path to merged SV solo datasets.}

\item{win_indel}{Numeric. Insertion and deletion error window.}

\item{win_inv_trans}{Numeric. Inversion and translocation error window.}

\item{perc_similarity}{Numeric . ThresholdPercentage similarity 
of the query SV and reference SV.}

\item{indelconf}{Numeric. Threshold for translocation Score.}

\item{returnMethod}{character. Choice between Text and DataFrame.}

\item{EnzymeType}{Character. Type of enzyme. Options Dual and DLE.}

\item{mergedFiles}{character. Path to the merged SV files.}

\item{buildSVInternalDB}{boolean. Checking whether the merged solo 
file database exist.}

\item{inputfmt}{character. Choice between Text and DataFrame.}

\item{path}{character. Path to the solo file database.}

\item{pattern}{character. pattern of the file names to merge.}
}
\value{
Text file or data frames containing internalFrequency data.
}
\description{
Calculates the internal frequencies of SV in internal cohorts
}
\examples{
path = "Z:/bionano/VSVM_Solo/UDN_Samples_Meged.txt"
smappath = "Z:\\\\bionano\\\\"
win_indel = 10000; win_inv_trans = 50000; perc_similarity = 0.5;
indelconf = 0.5; invconf = 0.01;transconf = 0.1
internalFrequency(mergedFiles , smappath , buildSVInternalDB=FALSE, 
win_indel, win_inv_trans, perc_similarity , indelconf, invconf , 
transconf,returnMethod="dataFrame")
}
