% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/Bed_SV_Comp_WIP.r
=======
% Please edit documentation in R/Bed_SV_Comp.r
>>>>>>> 04c1b43bc3aded9fbd47aace30093a5b16389a06
\name{compSmapbed}
\alias{compSmapbed}
\title{Extracts gene information from bed files}
\usage{
<<<<<<< HEAD
compSmapbed(
  smap,
  bed,
  inputfmtBed = c("BED", "BNBED"),
  EnzymeType = c("Dual", "DLE"),
  outpath,
  n = 3,
  returnMethod_bedcomp = c("Text", "dataFrame"),
  input_fmt_smap = c("Text", "dataFrame"),
  bperrorindel = 3000,
  bperrorinvtrans = 10000
)
=======
compSmapbed(smap, bed, inputfmtBed = c("BED", "BNBED"), outpath, n = 3,
  returnMethod_bedcomp = c("Text", "dataFrame"))
>>>>>>> 04c1b43bc3aded9fbd47aace30093a5b16389a06
}
\arguments{
\item{smap}{character. Path to SMAP file.}

\item{bed}{Text. Normal Bed files or Bionano Bed file.}

<<<<<<< HEAD
\item{inputfmtBed}{character Whether the bed input is UCSC bed or Bionano bed.
Note: extract in bed format to be read by bedsv:
awk '{print $1,$4,$5,$18,$7}' gencode.v19.annotation.gtf >HomoSapienGRCH19.bed}

\item{EnzymeType}{Character. Type of enzyme. Options Dual and DLE.}
=======
\item{inputfmtBed}{character Whether the bed input is UCSC bed or 
Bionano bed.
Note: extract in bed format to be read by bedsv:
awk '{print $1,$4,$5,$18,$7}' gencode.v19.annotation.gtf>HomoSapienGRCH19.bed}
>>>>>>> 04c1b43bc3aded9fbd47aace30093a5b16389a06

\item{outpath}{character Path for the output files.}

\item{n}{numeric Number of genes to report which are nearest to the
<<<<<<< HEAD
breakpoint. Default is        3.}

\item{returnMethod_bedcomp}{Character. Type of output Dataframe or in Text format.}

\item{bperrorindel}{Numeric. base pair error indel.}

\item{bperrorinvtrans}{Numeric. base pair error invtranslocation.}
}
\value{
Data Frame and Text file. Contains the smap with additional Gene Information.
=======
breakpoint. Default is   3.}

\item{returnMethod_bedcomp}{Character. Type of output Dataframe or in 
Text format.}
}
\value{
Data Frame and Text file. Contains the smap with additional 
Gene Information.
>>>>>>> 04c1b43bc3aded9fbd47aace30093a5b16389a06
}
\description{
Extracts gene information from bed files
}
\examples{
<<<<<<< HEAD
smapName="F1.1_GM24385_DLE-1_P_trio_hg19.smap"
smap = system.file("extdata", smapName, package="nanotatoR")
bedFile <- system.file("extdata", "Homo_sapiens.Hg19.bed", package="nanotatoR")
outpath <- system.file("extdata",    package="nanotatoR")
datcomp<-compSmapbed(smap, bed=bedFile, inputfmtBed =    "BED", outpath,
n = 3, returnMethod_bedcomp = c("dataFrame"), EnzymeType = "Dual")
=======
smapName="F1.1_TestSample1_solo_hg19.smap"
smap = system.file("extdata", smapName, package="nanotatoR")
bedFile <- system.file("extdata", "Homo_sapiens.Hg19_BN_bed.txt",
package="nanotatoR")
outpath <- system.file("extdata",  package="nanotatoR")
datcomp<-compSmapbed(smap, bed=bedFile, inputfmtBed =  "BNBED", n = 3, 
returnMethod_bedcomp = c("dataFrame"))
>>>>>>> 04c1b43bc3aded9fbd47aace30093a5b16389a06
datcomp[1,]
}
