% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DecipherExtraction.r
\name{Decipher_extraction}
\alias{Decipher_extraction}
\title{Frequency calculation of variants compared to DGV.}
\usage{
Decipher_extraction(
  decipherpath,
  smappath,
  smap,
  smap_data,
  input_fmt = c("Text", "dataFrame"),
  win_indel = 10000,
  perc_similarity = 0.5,
  returnMethod = c("Text", "dataFrame")
)
}
\arguments{
\item{smappath}{character. Path and file name for textfile.}

\item{smap_data}{Dataset containing smap data.}

\item{input_fmt}{character. Choice between text or data frame as
an input to the DGV frequency calculator.}

\item{returnMethod}{character. Choice between text or data frame as the output.}

\item{hgpath}{character. Path to Database of Genomic Variants (DGV)
Text file.}

\item{terms}{character. Single or Multiple Terms.}

\item{outpath}{character. Path where gene lists are saved.}

\item{thresh}{integer. Threshold for the number of terms sent to entrez.
Note if large lists are sent to ncbi, it might fail to get
processed. Default is 5.}
}
\value{
Text and character vector containg gene list and terms associated with them
        are stored as text files.
}
\description{
Frequency calculation of variants compared to DGV.
}
\examples{
\dontrun{
decipherpath="Z:/Suro/Annotator/Data/population_cnv.txt"
smappath="Z:/Suro/Annotator/Data/";
smap="F1.1_UDN287643_P_Q.S_VAP_SVmerge_trio_original2.txt";
win_indel=10000;win_inv_trans=50000;perc_similarity=0.5
DGV_extraction (hgpath, smappath, win_indel = 10000, win_inv_trans = 50000,
perc_similarity = 0.5,returnMethod="dataFrame")
}
}
