% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGV_extraction.r
\name{DGV_extraction}
\alias{DGV_extraction}
\title{Frequency calculation of variants compared to DGV.}
\usage{
DGV_extraction(hgpath, smappath, smap_data, input_fmt = c("Text",
  "DataFrame"), win_indel = 10000, win_inv_trans = 50000,
  perc_similarity = 0.5, returnMethod = c("Text", "dataFrame"))
}
\arguments{
\item{hgpath}{character. Path to Database of Genomic Variants (DGV)
Text file.}

\item{smappath}{character. Path and file name for textfile.}

\item{smap_data}{Dataset containing smap data.}

\item{input_fmt}{character. Choice between text or data frame as 
an input to the DGV frequency calculator.}

\item{returnMethod}{character. Choice between text or data frame as the output.}

\item{terms}{character. Single or Multiple Terms.}

\item{outpath}{character. Path where gene lists are saved.}

\item{thresh}{integer. Threshold for the number of terms sent to entrez.
Note if large lists are sent to ncbi, it might fail to get
processed. Default is 5.}
}
\value{
Text and character vector containg gene list and terms associated with them
        are stored as text files.
}
\description{
Frequency calculation of variants compared to DGV.
}
\examples{
\dontrun{
hgpath="C:\\\\Annotator\\\\Data\\\\GRCh37_hg19_variants_2016-05-15.txt";
smappath="Z:/Hayk's_Materials/Bionano/Projects/UDN/F1_UDN287643_Benic.Aria/";
smap="F1_UDN287643_P_Q.S_VAP_SVmerge_trio_access.txt";
win_indel=10000;win_inv_trans=50000;perc_similarity=0.5
DGV_extraction (hgpath, smappath, win_indel = 10000, win_inv_trans = 50000, 
perc_similarity = 0.5,returnMethod="dataFrame")
}
}
