% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrez_extract.r
\name{gene_list_generation}
\alias{gene_list_generation}
\title{Extracting genes for phenotype/diseases from NCBI.}
\usage{
gene_list_generation(method = c("Single", "Multiple", "Text"), termPath,
  term, outpath, thresh = 5, returnMethod_GeneList = c("Text",
  "dataFrame"))
}
\arguments{
\item{method}{character. Input Method for terms. Choices are
"Single","Multiple" and "Text".}

\item{termPath}{character. Path and file name for textfile.}

\item{term}{character. Single or Multiple Terms.}

\item{outpath}{character. Path where gene lists are saved.}

\item{thresh}{integer. Threshold for the number of terms sent to entrez.
Note if large lists are sent to ncbi, it might fail to get
processed. Default is 5.}

\item{returnMethod_GeneList}{Method of returning output. Options, Text or data.frame.}
}
\value{
Text files containg gene list and terms associated with them
        are stored as text files.
}
\description{
Extracting genes for phenotype/diseases from NCBI.
}
\examples{
terms="Muscle Weakness"
gene<-gene_list_generation(method="Single", term=terms,
  	returnMethod_GeneList="dataFrame")
}
