% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/gene_extract.r
=======
% Please edit documentation in R/entrez_extract.r
>>>>>>> 714868ed579a092f4cd6df397baa7f46eded618f
\name{gtr_gene}
\alias{gtr_gene}
\title{Extracting genes from gtr database NCBI.}
\usage{
<<<<<<< HEAD
gtr_gene(terms, gtr, url_gtr, downloadGTR = TRUE)
}
\arguments{
\item{terms}{Single or Multiple Terms.}

\item{gtr}{character gtr database location.}

\item{url_gtr}{character url for gtr database.}

\item{downloadGTR}{boolean If TRUE, download the gtr database.
Default FALSE.}
}
\value{
Dataframe returned containing gene lists in entrezid and Gene 
Symbols, and terms associated with it
=======
gtr_gene(terms)
}
\arguments{
\item{terms}{character. Single or Multiple Terms.}
}
\value{
Dataframe returned containing gene lists in entrezid and Gene 
        Symbols, and terms associated with it
>>>>>>> 714868ed579a092f4cd6df397baa7f46eded618f
}
\description{
Extracting genes from gtr database NCBI.
}
\examples{
<<<<<<< HEAD
terms="Muscle Weakness"
gtr = system.file("extdata", "gtrDatabase.txt", package="nanotatoR")
ge <- gtr_gene(terms = terms,gtr = gtr, downloadGTR = FALSE)
=======
\dontrun{
terms="Muscle Weakness"
gtr_gene(terms)
}
>>>>>>> 714868ed579a092f4cd6df397baa7f46eded618f
}
