% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrez_extract.r
\name{gtr_gene}
\alias{gtr_gene}
\title{Extracting genes from gtr database NCBI.}
\usage{
gtr_gene(terms)
}
\arguments{
\item{terms}{character. Single or Multiple Terms.}
}
\value{
Dataframe returned containing gene lists in entrezid and Gene 
        Symbols, and terms associated with it
}
\description{
Extracting genes from gtr database NCBI.
}
\examples{
\dontrun{
terms="Muscle Weakness"
gtr_gene(terms)
}
}
