% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_frequency.r
\name{internalFrequency}
\alias{internalFrequency}
\title{Calculates the internal frequencies of SV in internal cohorts}
\usage{
internalFrequency(mergedFiles, smappath, smapFileName,
  buildSVInternalDB = FALSE, smapdata, smapName, input_fmt = c("Text",
  "dataFrame"), path, pattern, outpath, win_indel = 10000,
  win_inv_trans = 50000, perc_similarity = 0.5, indelconf = 0.5,
  invconf = 0.01, transconf = 0.1, returnMethod = c("Text",
  "dataFrame"))
}
\arguments{
\item{mergedFiles}{character. Path to the merged SV files.}

\item{smappath}{character. path to the query smap file.}

\item{buildSVInternalDB}{boolean. Checking whether the merged solo 
file database exist.}

\item{path}{character. Path to the solo file database.}

\item{pattern}{character. pattern of the file names to merge.}

\item{outpath}{character. Path to merged SV solo datasets.}

\item{win_indel}{Numeric. Insertion and deletion error window.}

\item{win_inv_trans}{Numeric. Inversion and translocation error window.}

\item{perc_similarity}{Numeric . ThresholdPercentage similarity 
of the query SV and reference SV.}

\item{indelconf}{Numeric. Threshold for insertion and deletion confidence.}

\item{returnMethod}{character. Choice between Text and DataFrame.}

\item{smap}{character. File name for the smap}

\item{inputfmt}{character. Choice between Text and DataFrame.}

\item{indelconf}{Numeric. Threshold for inversion confidence.}

\item{indelconf}{Numeric. Threshold for translocation confidence.}
}
\value{
Text file or data frames containing internalFrequency data.
}
\description{
Calculates the internal frequencies of SV in internal cohorts
}
\examples{
\dontrun{
mergedFiles = "Z:/bionano/VSVM_Solo/UDN_Samples_Meged.txt"
smappath = "Z:\\\\bionano\\\\"
win_indel = 10000; win_inv_trans = 50000; perc_similarity = 0.5;
indelconf = 0.5; invconf = 0.01;transconf = 0.1
internalFrequency(mergedFiles , smappath , buildSVInternalDB=FALSE, 
win_indel, win_inv_trans, perc_similarity , indelconf, invconf , 
transconf,returnMethod="dataFrame")
}
}
