% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainAnnotator.r
\name{nanotatoR_main}
\alias{nanotatoR_main}
\title{Annotation of Bionano SV.}
\usage{
nanotatoR_main(smap, bed, inputfmtBed = c("BED", "BNBED"), n = 3,
  InternalDBpath, InternalDBpattern, dbOutput_Int, fname_Int, dbOutput_BN,
  fname_BN, buildSVInternalDB = FALSE, soloPath, solopattern,
  returnMethod_bedcomp = c("Text", "dataFrame"), mergedFiles_BN,
  win_indel_INF = 10000, win_inv_trans_INF = 50000,
  perc_similarity_INF = 0.5, indelconf = 0.5, invconf = 0.01,
  transconf = 0.1, returnMethod_DGV = c("Text", "dataFrame"), hgpath,
  win_indel_DGV = 10000, win_inv_trans_DGV = 50000,
  perc_similarity_DGV = 0.5, returnMethod_Internal = c("Text",
  "dataFrame"), input_fmt_DGV = c("Text", "dataFrame"),
  input_fmt_BN = c("Text", "dataFrame"), input_fmt_INF = c("Text",
  "dataFrame"), input_fmt_decipher = c("Text", "dataFrame"),
  input_fmt_svMap = c("Text", "dataFrame"), dat_geneList, decipherpath,
  input_fmt_geneList = c("Text", "dataFrame"),
  returnMethod_GeneList = c("Text", "dataFrame"),
  buildBNInternalDB = FALSE, returnMethod_BNCohort = c("Text",
  "dataFrame"), mergedFiles_INF, returnMethod_decipher = c("Text",
  "dataFrame"), method_entrez = c("Single", "Multiple", "Text"),
  smapName, termPath, term, thresh = 5, limsize = 1000,
  win_indel_parents = 5000, win_inv_trans_parents = 40000, outpath,
  outputFilename = "", RZIPpath)
}
\arguments{
\item{smap}{character. Path to SMAP file.}

\item{bed}{Text Choice between UCSC bed or Bionano bed.}

\item{inputfmtBed}{character. Choice between Text and DataFrame
as input for bed file.}

\item{n}{numeric Number of genes to report which are nearest to 
the breakpoint.
Default is   3.}

\item{InternalDBpath}{character. Path to the BNFile file database.}

\item{InternalDBpattern}{character. pattern of the BNFile names to merge.}

\item{dbOutput_Int}{character. Output of solo bionano data.}

\item{fname_Int}{character. Filename in case dbOutput_Int = Text.}

\item{dbOutput_BN}{character. Output of merged bionano data.}

\item{fname_BN}{character. Filename in case dbOutput_BN = Text.}

\item{buildSVInternalDB}{boolean. Checking whether the merged solo
file database exist or you need to build it. Default= TRUE.}

\item{soloPath}{character. Path to the solo file database.}

\item{solopattern}{character. pattern of the file names to merge.}

\item{returnMethod_bedcomp}{character. Return Methods from the compSmapbed
function, choice between Text and Dataframe.}

\item{mergedFiles_BN}{character. Path to the merged BN SV files.}

\item{win_indel_INF}{Numeric. Insertion and deletion error window.}

\item{win_inv_trans_INF}{Numeric. Inversion and translocation error window.}

\item{perc_similarity_INF}{Numeric . ThresholdPercentage similarity
of the query SV and reference SV.}

\item{indelconf}{Numeric. Threshold for insertion and deletion confidence.}

\item{invconf}{Numeric. Threshold for inversion confidence.}

\item{transconf}{Numeric. Threshold for translocation confidence.}

\item{returnMethod_DGV}{character. Return Methods from 
the DGV_extraction function, choice between
Text and Dataframe.}

\item{hgpath}{character. Path to Database of Genomic Variants (DGV)
Text file.}

\item{win_indel_DGV}{Numeric. Insertion and deletion error window.}

\item{win_inv_trans_DGV}{Numeric. Inversion and translocation error window.}

\item{perc_similarity_DGV}{Numeric . ThresholdPercentage similarity
of the query SV and reference SV.}

\item{returnMethod_Internal}{character. Return Methods 
from the internalFrequency function, choice between
Text and Dataframe.}

\item{input_fmt_DGV}{character. Choice between Text and DataFrame
for input to DGV.}

\item{input_fmt_BN}{character. Choice of Bionano dataset input
Text or Dataframe.}

\item{input_fmt_INF}{character. Choice between Text and DataFrame
for input to INF.}

\item{input_fmt_decipher}{character. Choice of gene list input
Text or Dataframe.}

\item{input_fmt_svMap}{character. Choice of SVmap input for final step
Text or Dataframe.}

\item{dat_geneList}{Dataframe Input data containing geneList data.}

\item{decipherpath}{character. Path to DECIPHER. Text file.}

\item{input_fmt_geneList}{character. Choice of gene list input
Text or Dataframe.}

\item{returnMethod_GeneList}{character. Return Methods from the 
gene_list_generation
function, choice between Text and Dataframe.}

\item{buildBNInternalDB}{boolean. Checking whether the merged Bionano
file database exist or you need to build it. Default= TRUE.}

\item{returnMethod_BNCohort}{character. Return 
Methods from the Bionano function, choice between
Text and Dataframe.}

\item{mergedFiles_INF}{character. Path to the merged BN SV files.}

\item{returnMethod_decipher}{character.
Return Methods from the decipher Frequency function, choice between
Text and Dataframe.}

\item{method_entrez}{character. Input Method for terms for entrez. Choices 
are "Single","Multiple" and "Text".}

\item{smapName}{character. Name of the smap file.}

\item{termPath}{character. Path and file name for textfile for terms.}

\item{term}{character. Single or Multiple Terms as vectord.}

\item{thresh}{integer. Threshold for the number of terms sent to entrez.
Note if large lists are sent to ncbi, it might fail to get
processed. Default is 5.}

\item{limsize}{Numeric. Minimum size of SV that can be determined 
accurately by the Bionano SV caller. Default 1000.}

\item{win_indel_parents}{Numeric. Insertion and deletion error window to 
determine zygosity in case of parents. Default 5000.}

\item{win_inv_trans_parents}{Numeric. Inversion and translocation error 
window to determine zygosity in case of parents. Default 40000.}

\item{outpath}{Character Directory to the output file.}

\item{outputFilename}{Character Output filename for the annotated smap.}

\item{RZIPpath}{Character. Path for the Rtools zip.exe}
}
\value{
Excel file containing the annotated SV map, tabs divided based on
type of SVs.
}
\description{
Annotation of Bionano SV.
}
