% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bed_SV_Comp.r
\name{nonOverlapGenes}
\alias{nonOverlapGenes}
\title{Calculates Genes that are near to the SV region}
\usage{
nonOverlapGenes(bed, chrom, startpos, endpos, svid, SVTyp, n = 3)
}
\arguments{
\item{bed}{Text Bionano Bed file.}

\item{chrom}{character SVmap chromosome.}

\item{startpos}{numeric starting position of the breakpoints.}

\item{endpos}{numeric end position of the breakpoints.}

\item{svid}{numeric Structural variant identifier (Bionano generated).}

\item{n}{numeric Number of genes to report which are nearest to the breakpoint.
Default is      3.}
}
\value{
Data Frame. Contains the SVID,Gene name,strand information and 
Distance from the SV covered.
}
\description{
Calculates Genes that are near to the SV region
}
\examples{
\dontrun{
smap="C:\\\\Annotator\\\\Data\\\\F1.1_UDN287643_P_Q.S_VAP_SVmerge_trio_original.txt"
bedFile = "C:\\\\Annotator\\\\Data\\\\Homo_sapiens.GRCH19.bed"
bed<-buildrunBNBedFiles(bedFile,returnMethod = "dataFrame")
smap<-readSMap(smap)
chrom<-smap$RefcontigID1
startpos<-smap$RefStartPos
endpos<-smap$RefEndPos
svid<-smap$SVIndex
n<-3
nonOverlapGenes(bed, chrom, startpos, endpos, svid,n)
}
}
