% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bed_SV_Comp.r
\name{nonOverlapGenes}
\alias{nonOverlapGenes}
\title{Calculates Genes that are near to the SV region}
\usage{
nonOverlapGenes(bed, chrom, startpos, endpos, svid, n = 3)
}
\arguments{
\item{bed}{Text Bionano Bed file.}

\item{chrom}{character SVmap chromosome.}

\item{startpos}{numeric starting position of the breakpoints.}

\item{endpos}{numeric end position of the breakpoints.}

\item{svid}{numeric Structural variant identifier (Bionano generated).}

\item{n}{numeric Number of genes to report which are nearest to the breakpoint.
Default is 3.}
}
\value{
Data Frame. Contains the SVID,Gene name,strand information and
Distance from the SV covered.
}
\description{
Calculates Genes that are near to the SV region
}
\examples{
smapName="F1.1_GM24385_DLE-1_P_trio_hg19.smap"
smap = system.file("extdata", smapName, package="nanotatoR")
bedFile <- system.file("extdata", "Homo_sapiens.Hg19.bed", package="nanotatoR")
bed<-buildrunBNBedFiles(bedFile,returnMethod="dataFrame")
smap<-readSMap(smap)
chrom<-smap$RefcontigID1
startpos<-smap$RefStartPos
endpos<-smap$RefEndPos
if (length(grep("SVIndex",names(smap)))>0){
   svid <- smap$SVIndex
 }else{
	svid <- smap$SmapEntryID
	}
n<-3
nonOverlapGenes(bed, chrom, startpos, endpos, svid,n)
}
