% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASEQ_Analysis_Duo_Trio_SVMerge_SE.r
\name{nonOverlapRNAseq}
\alias{nonOverlapRNAseq}
\title{Extract Read counts for genes that are near SVs.}
\usage{
nonOverlapRNAseq(
  gnsNonOverlap,
  SVID,
  RNASeqData,
  pattern_Proband = NA,
  pattern_Mother = NA,
  pattern_Father = NA
)
}
\arguments{
\item{gnsNonOverlap}{character. genes that are upstream 
and/or downstream of SV.}

\item{SVID}{character. ID of the SVs.}

\item{RNASeqData}{character. Expression of the genes.}

\item{pattern_Proband}{character. Pattern to identify the proband reads.}

\item{pattern_Mother}{character. Pattern to identify the mother reads.}

\item{pattern_Father}{character. Pattern to identify the father reads.}
}
\value{
Text or Dataframe containing TPM read counts of genes in the family.
}
\description{
Extract Read counts for genes that are near SVs.
}
\examples{
RNASeqDir = system.file("extdata", package="nanotatoR")
returnMethod="dataFrame"
datRNASeq <- RNAseqcombine(RNASeqDir = RNASeqDir,
returnMethod = returnMethod)
gnsNonOverlap <- c("DDX11L1", "MIR1302-2HG", "OR4G4P")
SVID = 397
datgnnonovrlap <- nonOverlapRNAseq(gnsNonOverlap = gnsNonOverlap, 
SVID = SVID, RNASeqData = datRNASeq,
pattern_Proband = "*_P_*")
}
