% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASEQ_Analysis_solo.r
\name{nonOverlapRNAseq_solo}
\alias{nonOverlapRNAseq_solo}
\title{Annotating the Non-Overlapping genes with RNAseq expression}
\usage{
nonOverlapRNAseq_solo(gnsNonOverlap, SVID, RNASeqData, pattern_Proband = NA)
}
\arguments{
\item{gnsNonOverlap}{character. Vector containing non-overlapping genes.}

\item{SVID}{character. SV Index ID.}

\item{RNASeqData}{dataFrame. RNAseq data with gene names.}

\item{pattern_Proband}{character. Pattern for proband.}
}
\value{
Dataframe containing TPM read counts of overlapping genes.
}
\description{
Annotating the Non-Overlapping genes with RNAseq expression
}
\examples{
RNASeqDir="Z:/Suro/Annotator/Data/RNAseq"
returnMethod="dataFrame"
RNAseqcombine(RNASeqDir,returnMethod)
}
