% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_extract.r
\name{omim_gene}
\alias{omim_gene}
\title{Extracting genes from OMIM database NCBI.}
\usage{
omim_gene(terms, omim)
}
\arguments{
\item{terms}{character Single or Multiple Terms.}

\item{omim}{character omim database location.}
}
\value{
Dataframe returned containing gene lists in entrezid and Gene 
Symbols, and terms associated with it
}
\description{
Extracting genes from OMIM database NCBI.
}
\examples{
terms="Muscle Weakness"
omim = system.file("extdata", "mim2gene.txt", package="nanotatoR")
ge <- omim_gene(terms = terms, omim = omim)
}
