% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/gene_extract.r
=======
% Please edit documentation in R/entrez_extract.r
>>>>>>> 714868ed579a092f4cd6df397baa7f46eded618f
\name{omim_gene}
\alias{omim_gene}
\title{Extracting genes from OMIM database NCBI.}
\usage{
<<<<<<< HEAD
omim_gene(terms, omim)
}
\arguments{
\item{terms}{character Single or Multiple Terms.}

\item{omim}{character omim database location.}
}
\value{
Dataframe returned containing gene lists in entrezid and Gene 
Symbols, and terms associated with it
=======
omim_gene(terms)
}
\arguments{
\item{terms}{character. Single or Multiple Terms.}
}
\value{
Dataframe returned containing gene lists in entrezid and Gene 
        Symbols, and terms associated with it
>>>>>>> 714868ed579a092f4cd6df397baa7f46eded618f
}
\description{
Extracting genes from OMIM database NCBI.
}
\examples{
<<<<<<< HEAD
terms="Muscle Weakness"
omim = system.file("extdata", "mim2gene.txt", package="nanotatoR")
ge <- omim_gene(terms = terms, omim = omim)
=======
\dontrun{
terms="Muscle Weakness"
omim_gene(terms)
}
>>>>>>> 714868ed579a092f4cd6df397baa7f46eded618f
}
