% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrez_extract.r
\name{omim_gene}
\alias{omim_gene}
\title{Extracting genes from OMIM database NCBI.}
\usage{
<<<<<<< HEAD
omim_gene(terms, omim)
}
\arguments{
\item{terms}{character Single or Multiple Terms.}

\item{omim}{character omim database location.}
=======
omim_gene(terms)
}
\arguments{
\item{terms}{Single or Multiple Terms.}
>>>>>>> 04c1b43bc3aded9fbd47aace30093a5b16389a06
}
\value{
Dataframe returned containing gene lists in entrezid and Gene 
Symbols, and terms associated with it
}
\description{
Extracting genes from OMIM database NCBI.
}
\examples{
terms="Muscle Weakness"
<<<<<<< HEAD
ge <- omim_gene(terms, omim = "Y:/Suro/nanotatoRDatabases/mim2gene.txt")
=======
ge <- omim_gene(terms)
>>>>>>> 04c1b43bc3aded9fbd47aace30093a5b16389a06
}
