% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/Bed_SV_Comp_WIP.r
=======
% Please edit documentation in R/Bed_SV_Comp.r
>>>>>>> 04c1b43bc3aded9fbd47aace30093a5b16389a06
\name{overlapGenes}
\alias{overlapGenes}
\title{Calculates Genes that overlap the SV region}
\usage{
<<<<<<< HEAD
overlapGenes(
  bed,
  chrom,
  startpos,
  endpos,
  svid,
  chrom2,
  SVTyp,
  bperrorindel = 3000,
  bperrorinvtrans = 10000
)
=======
overlapGenes(bed, chrom, startpos, endpos, svid)
>>>>>>> 04c1b43bc3aded9fbd47aace30093a5b16389a06
}
\arguments{
\item{bed}{Text Bionano Bed file.}

\item{chrom}{character SVmap chromosome.}

\item{startpos}{numeric starting position of the breakpoints.}

\item{endpos}{numeric end position of the breakpoints.}

\item{svid}{numeric Structural variant identifier (Bionano generated).}
}
\value{
Data Frame. Contains the SVID,Gene name,strand information and
percentage of SV covered.
}
\description{
Calculates Genes that overlap the SV region
}
\examples{
<<<<<<< HEAD
smapName="F1.1_GM24385_DLE-1_P_trio_hg19.smap"
smap = system.file("extdata", smapName, package="nanotatoR")
bedFile <- system.file("extdata", "Homo_sapiens.Hg19.bed", 
       package="nanotatoR")
=======
smapName="F1.1_TestSample1_solo_hg19.smap"
smap = system.file("extdata", smapName, package="nanotatoR")
bedFile <- system.file("extdata", "Homo_sapiens.Hg19.bed", 
 package="nanotatoR")
>>>>>>> 04c1b43bc3aded9fbd47aace30093a5b16389a06
bed<-buildrunBNBedFiles(bedFile,returnMethod="dataFrame")
smap<-readSMap(smap)
chrom<-smap$RefcontigID1
startpos<-smap$RefStartPos
endpos<-smap$RefEndPos
if (length(grep("SVIndex",names(smap)))>0){
<<<<<<< HEAD
       svid <- smap$SVIndex
   }else{
    svid <- smap$SmapEntryID
    }
=======
   svid <- smap$SVIndex
 }else{
 svid <- smap$SmapEntryID
 }
>>>>>>> 04c1b43bc3aded9fbd47aace30093a5b16389a06
overlapGenes(bed, chrom, startpos, endpos, svid)
}
