% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bed_SV_Comp.r
\name{overlapGenes}
\alias{overlapGenes}
\title{Calculates Genes that overlap the SV region}
\usage{
overlapGenes(bed, chrom, startpos, endpos, svid)
}
\arguments{
\item{bed}{Text Bionano Bed file.}

\item{chrom}{character SVmap chromosome.}

\item{startpos}{numeric starting position of the breakpoints.}

\item{endpos}{numeric end position of the breakpoints.}

\item{svid}{numeric Structural variant identifier (Bionano generated).}
}
\value{
Data Frame. Contains the SVID,Gene name,strand information and
percentage of SV covered.
}
\description{
Calculates Genes that overlap the SV region
}
\examples{
smapName="F1.1_TestSample1_solo_hg19.smap"
smap = system.file("extdata", smapName, package="nanotatoR")
bedFile <- system.file("extdata", "Homo_sapiens.Hg19.bed", 
 package="nanotatoR")
bed<-buildrunBNBedFiles(bedFile,returnMethod="dataFrame")
smap<-readSMap(smap)
chrom<-smap$RefcontigID1
startpos<-smap$RefStartPos
endpos<-smap$RefEndPos
if (length(grep("SVIndex",names(smap)))>0){
   svid <- smap$SVIndex
 }else{
 svid <- smap$SmapEntryID
 }
overlapGenes(bed, chrom, startpos, endpos, svid)
}
