% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bed_SV_Comp.r
\name{overlapGenes}
\alias{overlapGenes}
\title{Calculates Genes that overlap the SV region}
\usage{
overlapGenes(bed, chrom, startpos, endpos, svid)
}
\arguments{
\item{bed}{Text Bionano Bed file.}

\item{chrom}{character SVmap chromosome.}

\item{startpos}{numeric starting position of the breakpoints.}

\item{endpos}{numeric end position of the breakpoints.}

\item{svid}{numeric Structural variant identifier (Bionano generated).}
}
\value{
Data Frame. Contains the SVID,Gene name,strand information and
percentage of SV covered.
}
\description{
Calculates Genes that overlap the SV region
}
\examples{
smap="C:\\\\nanotatoR\\\\Data\\\\F1.1_UDN287643_P_Q.S_VAP_SVmerge_trio_original.txt"
bedFile = "C:\\\\nanotatoR\\\\Data\\\\Homo_sapiens.GRCH19.bed"
bed<-buildrunBNBedFiles(bedFile,returnMethod = "dataFrame")
smap<-readSMap(smap)
chrom<-smap$RefcontigID1
startpos<-smap$RefStartPos
endpos<-smap$RefEndPos
svid<-smap$SVIndex
overlapGenes(bed, chrom, startpos, endpos, svid)
}
