% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/OverlappingGenes.r
=======
% Please edit documentation in R/Bed_SV_Comp.r
>>>>>>> 714868ed579a092f4cd6df397baa7f46eded618f
\name{readBNBedFiles}
\alias{readBNBedFiles}
\title{Reads Bionano Bedfiles}
\usage{
readBNBedFiles(BNFile)
}
\arguments{
\item{BNFile}{character. Path to Bionano Bed File.}
}
\value{
Data Frame Contains the gene information.
}
\description{
Reads Bionano Bedfiles
}
\examples{
<<<<<<< HEAD
BNFile <- system.file("extdata", "HomoSapienGRCH19_lift37_BN.bed", package="nanotatoR")
bed<-readBNBedFiles(BNFile)
=======
\dontrun{
}
>>>>>>> 714868ed579a092f4cd6df397baa7f46eded618f
}
