% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bed_SV_Comp.r
\name{readBNBedFiles}
\alias{readBNBedFiles}
\title{Reads Bionano Bedfiles}
\usage{
readBNBedFiles(BNFile)
}
\arguments{
\item{BNFile}{character. Path to Bionano Bed File.}
}
\value{
Data Frame Contains the gene information.
}
\description{
Reads Bionano Bedfiles
}
\examples{
\dontrun{
}
}
