% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/Bed_SV_Comp_WIP.r
=======
% Please edit documentation in R/Bed_SV_Comp.r
>>>>>>> 04c1b43bc3aded9fbd47aace30093a5b16389a06
\name{readSMap}
\alias{readSMap}
\title{Reads SMAP files to extract information}
\usage{
<<<<<<< HEAD
readSMap(smap, input_fmt_smap = c("Text", "dataFrame"))
=======
readSMap(smap)
>>>>>>> 04c1b43bc3aded9fbd47aace30093a5b16389a06
}
\arguments{
\item{smap}{character. Path to SMAP file.}
}
\value{
Data Frame or text file. Contains the SMAP information.
}
\description{
Reads SMAP files to extract information
}
\examples{
<<<<<<< HEAD
smapName="F1.1_GM24385_DLE-1_P_trio_hg19.smap"
=======
smapName="F1.1_TestSample1_solo_hg19.smap"
>>>>>>> 04c1b43bc3aded9fbd47aace30093a5b16389a06
smap = system.file("extdata", smapName, package="nanotatoR")
readSMap(smap)
}
