% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/OverlappingGenes.r
\name{readSMap}
\alias{readSMap}
\title{Reads SMAP files to extract information from SVMerge}
\usage{
readSMap(smap, smapdata, input_fmt_smap = c("Text", "dataFrame"))
}
\arguments{
\item{smap}{character. Path to SMAP file.}

\item{smapdata}{dataframe. variable for smap dataset.}

\item{input_fmt_smap}{character. input format for smap
text or dataframe.}
=======
% Please edit documentation in R/Bed_SV_Comp.r
\name{readSMap}
\alias{readSMap}
\title{Reads SMAP files to extract information}
\usage{
readSMap(smap)
}
\arguments{
\item{smap}{character. Path to SMAP file.}
>>>>>>> 714868ed579a092f4cd6df397baa7f46eded618f
}
\value{
Data Frame or text file. Contains the SMAP information.
}
\description{
<<<<<<< HEAD
Reads SMAP files to extract information from SVMerge
}
\examples{
smapName="NA12878_Q.S_VAP_SVmerge_solo5.txt"
smap = system.file("extdata", smapName, package="nanotatoR")
readSMap(smap, input_fmt_smap = "Text")
=======
Reads SMAP files to extract information
}
\examples{
\dontrun{
smap = "C:\\\\Annotator\\\\Data\\\\F1_UDN287643_P_Q.S_VAP_SVmerge_trio.txt"
readSMap(smap)
}
>>>>>>> 714868ed579a092f4cd6df397baa7f46eded618f
}
