% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bed_SV_Comp.r
\name{readSMap}
\alias{readSMap}
\title{Reads SMAP files to extract information}
\usage{
readSMap(smap)
}
\arguments{
\item{smap}{character. Path to SMAP file.}
}
\value{
Data Frame or text file. Contains the SMAP information.
}
\description{
Reads SMAP files to extract information
}
\examples{
smapName="F1.1_TestSample1_solo_hg19.smap"
smap = system.file("extdata", smapName, package="nanotatoR")
readSMap(smap)
}
