% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OverlappingGenes.r
\name{readSMap}
\alias{readSMap}
\title{Reads SMAP files to extract information from SVMerge}
\usage{
readSMap(smap, smapdata, input_fmt_smap = c("Text", "dataFrame"))
}
\arguments{
\item{smap}{character. Path to SMAP file.}

\item{smapdata}{dataframe. variable for smap dataset.}

\item{input_fmt_smap}{character. input format for smap
text or dataframe.}
}
\value{
Data Frame or text file. Contains the SMAP information.
}
\description{
Reads SMAP files to extract information from SVMerge
}
\examples{
smapName="NA12878_Q.S_VAP_SVmerge_solo5.txt"
smap = system.file("extdata", smapName, package="nanotatoR")
readSMap(smap, input_fmt_smap = "Text")
}
