% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bed_SV_Comp_WIP.r
\name{readSMap_DLE}
\alias{readSMap_DLE}
\title{Reads DLE SMAP files to extract information}
\usage{
readSMap_DLE(smap, input_fmt_smap = "Text")
}
\arguments{
\item{smap}{character. Path to SMAP file.}
}
\value{
Data Frame or text file. Contains the SMAP information.
}
\description{
Reads DLE SMAP files to extract information
}
\examples{
smapName="F1.1_GM24385_DLE-1_P_trio_hg19.smap"
smap = system.file("extdata", smapName, package="nanotatoR")
readSMap(smap)
}
