% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterread.r
\name{run_bionano_filter}
\alias{run_bionano_filter}
\title{Getting the data from annotated smaps to extract SV 
information based on type of variants.}
\usage{
run_bionano_filter(input_fmt_geneList = c("Text", "dataFrame"),
  input_fmt_svMap = c("Text", "dataFrame"), SVFile = NULL, svData,
  dat_geneList, fileName, outpath, outputFilename = "")
}
\arguments{
\item{input_fmt_geneList}{character. Choice of gene list input 
Text or Dataframe.}

\item{input_fmt_svMap}{character. Choice of gene list input 
Text or Dataframe.}

\item{SVFile}{character. SV file name.}

\item{svData}{Dataframe Input data containing SV data.}

\item{dat_geneList}{Dataframe Input data containing geneList data.}

\item{fileName}{Character Name of file containing Gene List data.}

\item{outpath}{Character Directory to the output file.}

\item{outFileName}{Character Output filename.}
}
\value{
Excel file containing the annotated SV map, tabs divided based on
type of SVs.
}
\description{
Getting the data from annotated smaps to extract SV 
information based on type of variants.
}
\examples{
path="Z:/Hayk's_Materials/Bionano/Projects/UDN/F1_UDN287643_Benic.Aria"
SVFile="F1.1_UDN287643_P_Q.S_VAP_SVmerge_trio_DGV_Int.txt"
fileName="F1.1_UDN287643_P_Q.S_VAP_SVmerge_trio_DGV.txt"
run_bionano_filter(path,SVFile,fileName)
}
