% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterread.r
\name{run_bionano_filter}
\alias{run_bionano_filter}
\title{Getting the data from annotated smaps to extract SV 
information based on type of variants.}
\usage{
run_bionano_filter(input_fmt_geneList = c("Text", "dataFrame"),
  input_fmt_svMap = c("Text", "dataFrame"), SVFile = NULL, svData,
  dat_geneList, fileName, outpath, outputFilename = "", RZIPpath = "")
}
\arguments{
\item{input_fmt_geneList}{character. Choice of gene list input 
Text or Dataframe.}

\item{input_fmt_svMap}{character. Choice of gene list input 
Text or Dataframe.}

\item{SVFile}{character. SV file name.}

\item{svData}{Dataframe Input data containing SV data.}

\item{dat_geneList}{Dataframe Input data containing geneList data.}

\item{fileName}{Character Name of file containing Gene List data.}

\item{outpath}{Character Directory to the output file.}

\item{outputFilename}{Output filename.}

\item{RZIPpath}{Character. Path for the Rtools zip.exe}
}
\value{
Excel file containing the annotated SV map, tabs divided based on
type of SVs.
}
\description{
Getting the data from annotated smaps to extract SV 
information based on type of variants.
}
\examples{
\dontrun{
smapName="F1.1_GM24385_DLE-1_P_trio_hg19.smap"
smappath = system.file("extdata", smapName, package="nanotatoR")
terms="Muscle Weakness"
gene<-gene_list_generation(method="Single", term=terms, 
  	returnMethod_GeneList="dataFrame")
run_bionano_filter(SVFile=smappath,fileName,input_fmt_geneList="dataFrame",
      	input_fmt_svMap="Text",RtoolsZIPpath="")
}
}
