% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGV_extraction.r
\name{UniqueSample}
\alias{UniqueSample}
\title{Calculating number of samples.}
\usage{
UniqueSample(samp)
}
\arguments{
\item{samp}{character. Unique samples}
}
\description{
Calculating number of samples.
}
\examples{
hgpath=system.file("extdata", "GRCh37_hg19_variants_2016-05-15.txt", package="nanotatoR")
r <- read.table(paste(hgpath), header = TRUE, sep = "\\t")
samp <- as.character(unique(r$samples))
UniqueSample(samp)
}
